### 2023-02-21 #575 @bnicenboim
 submission: pangoling: Access to word predictability using large language (transformer) models
Submitting Author Name: Bruno Nicenboim
Submitting Author Github Handle: @bnicenboim
Repository:  https://github.com/bnicenboim/pangoling
Version submitted: 0.0.0.9005
Submission type: Standard

Version accepted: TBD
---

```
Package: pangoling
Title: Access to Large Language Model Predictions
Version: 0.0.0.9005
    person("Bruno", "Nicenboim",
    email = "bruno.nicenboim@gmail.com",
    role = c( "aut","cre"),
    person("Chris", "Emmerly", role = "ctb"),
    person("Giovanni", "Cassani", role = "ctb"))
Description: Access to word predictability using large language (transformer) models.
LazyData: false
Config/reticulate:
  list(
    packages = list(
      list(package = "torch"),
      list(package = "transformers")
    )
  )
    memoise,
    reticulate,
    tidyselect,
    tidytable (>= 0.7.2),
    cachem
    tictoc,
    covr,
    spelling
StagedInstall: yes

```

The package is built on top of the python package `transformers`, and it offers some basic functionality for text analysis, including tokenization and perplexity calculation. Crucially `pangoling` also offers word predictability, which is widely used as a predictor in psycho and neurolinguistics, and it's  not trivial to obtain.  Also `transformers` works with "tokens" rather than "words", and then pangoling takes cares of the mapping between the tokens to the target words (or even phrases).

This is mostly for psycho/neuro/- linguists that use word predictability as a predictor in their research, such as in ERP/EEG and reading studies.

Another R package that acts as a wrapper for [`transformers`](https://pypi.org/project/transformers/) is [`text`](https://r-text.org//) However, `text` is more general, and its focus  is on Natural Language Processing and Machine Learning. `pangoling` is much more specific and the focus is on measures used as predictors in analyses of data from experiments, rather than NLP.  `text` doesn't allow for generating pangoling output in a straightforward way and in fact, I'm not sure if it's even possible to get token probabilities from `text` since it seems more limited than the python package `transformers`.

NA

#573

`pkgcheck` fails only because of the use of `<<-`. But this is done in `.OnLoad` as recommended by reticulate. Also see [this issue](https://github.com/ropensci-review-tools/pkgcheck/issues/167) .

- [x] has a CRAN and OSI accepted license.  **-> I think so, it has an MIT license**
