### 2023-02-21 #575 @bnicenboim
Hi,
" Package coverage is 0.9% (should be at least 75%)." -> this seems to be an error.
See that the coverage is [94%](https://app.codecov.io/gh/bnicenboim/pangoling?branch=main).
Maybe it is because the tests skip almost everything, if there is no python installed? (As recommended by reticulate).

Also when I run `pkgcheck::pkgcheck()` I get the following:

✔ Package coverage is 94.8%.
