### 2023-02-26 #575 @maurolepore
I guess authors typically are guided by their knowledge of their intended audience. But for inspiration see how [editors look for reviewers](https://devguide.ropensci.org/editorguide.html#where-to-look-for-reviewers). Editors have access to a private airtable database, but often we look elsewhere.
