+++ man/masked_tokens_tbl.Rd
@@ -15,7 +15,7 @@ masked_tokens_tbl(
 \arguments{
 \item{masked_sentences}{Masked sentences.}
 
-\item{model}{Name of a pre-trained model.}
+\item{model}{Name of a pre-trained model or folder.}
 
 \item{add_special_tokens}{Whether to include special tokens. It has the
 same default as the
