+++ man/masked_preload.Rd
@@ -12,7 +12,7 @@ masked_preload(
 )
 }
 \arguments{
-\item{model}{Name of a pre-trained model.}
+\item{model}{Name of a pre-trained model or folder.}
 
 \item{add_special_tokens}{Whether to include special tokens. It has the
 same default as the
