+++ man/masked_config.Rd
@@ -10,7 +10,7 @@ masked_config(
 )
 }
 \arguments{
-\item{model}{Name of a pre-trained model.}
+\item{model}{Name of a pre-trained model or folder.}
 
 \item{config_model}{List with other arguments that control how the
 model from Hugging Face is accessed.}
