+++ man/tokenize_lst.Rd
@@ -14,7 +14,7 @@ tokenize_lst(
 \arguments{
 \item{x}{Strings or token ids.}
 
-\item{model}{Name of a pre-trained model.}
+\item{model}{Name of a pre-trained model or folder.}
 
 \item{add_special_tokens}{Whether to include special tokens. It has the
 same default as the
