+++ NEWS.md
@@ -19,6 +19,7 @@
 * Strings with no tokens no longer throw errors. 
 * Requires correct version of R. 
 
-# pangoling 0.0.0.9005
+# pangoling 0.0.0.9006
 * `causal_lp` get a `l_contexts` argument
 * checkpoints work for causal models (not yet for masked models)
+* Ropensci badge added.
