### 2023-05-30 #575 @lisalevinson
Thanks so much for your work on this package @bnicenboim! Sorry this review is a bit later than expected. I have made a lot of comments in my review on my personal user interface/workflow preferences, but the package is already very useful “as is”. Let me know if you have any questions, or would like any help with some of the vignette/documentation suggestions I have made - I’m happy to help, though I can’t promise a speedy turnaround.

- [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

I have no prior personal or working relationship with the package creator.

The Rmd vignettes exist (and are viewable as rendered on the package website), but they did not seem to install properly, either from the local package folder or from GitHub, even when building vignettes is set to true. I believe this may be related to an issue that occurs due to the `inst/docs` folder being deleted on build (there isn’t such a folder in the repo). Running `vignette(package = "pangoling")` yields “no vignettes found” for me.

Examples run, but one has a typo, described below.

I’m not experienced with testing, but there were some failures and warnings noted below. There did not seem to be many tests.

- [x] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

Thanks so much for creating this package - I demo’d it to my students when it first came out to show them ways to do “Python” things within the R environment, and it was a great example for that. I also am very grateful to have a tool that would allow me to more easily carry out these kinds of analyses from R, rather than passing data over from Python projects (as I had previously been doing). It seems similar in spirit to the Python package [minicons](https://github.com/kanishkamisra/minicons).

### Further Comments on Documentation

I did not have a “fresh” system without Python and `reticulate` already setup to test on. I’m not sure how automated it is in current workflows to set that up, and there isn’t any specific guidance within the package. If the target population is (psycho)linguists who aren’t familiar with Python, and might not even have Python installed, then they would probably need more guidance through this process.

Given the variety of inputs and outputs, it would be very helpful in the documentation/vignettes to have a table summarizing these for each function so it’s easier to tell which one you need for which purpose.

### Workflow

One general comment is that the functions could be a little easier to use for the output that I am usually needing to generate for analyses. In my work, I’m typically gathering probabilities/surprisals for all of the words in a sentence or passage, not just one target word. Thus the Python scripts I have been using are based on specifying a complete sentence (“The apple fell far from the tree.”) and either specifying a target word position or getting probabilities across all of the word positions. This is awkward to do with the `pangoling` functions. It is a bit more straightforward for the causal model functions, but since the masked model functions require arguments for “pre”, “target”, and “post”, it is more cumbersome. I would prefer an argument structure more like the following (starting counts at 1 as “native” in R):

    causal_lp(s = "The apple doesn't fall far from the tree.", target = 8)
    masked_lp(s = "The apple doesn't fall far from the tree.", mask = 2, fill = c("apple", "pear"))

The default “fill” for `masked_lp` could be the word present in the original sentence, so that the following would use the mask “banana”, and would be very similar to running the `causal_lp` function:

    masked_lp(s = "The banana doesn't fall far from the tree.", mask = 2)

Perhaps functions like these could be separate ones from those already provided.

### Reporting and Documentation of Probabilities

Based on some independent testing, it seems that the log probabilities are natural log, but I couldn’t find that documented anywhere. It is fairly common to use base 2 log for surprisals (in bits), so which log is used should at least be indicated clearly. It would be great to make this an argument which can be specified, as `base =` for the various functions.

On this note, it seems that most folks use surprisals in the psycholinguistic literature, and they are a bit easier to interpret than negative probabilities. It would be nice to offer a “surprisal” option, which would usually be negative log2 probability.

### Output Formats

The outputs of the functions are in a variety of structures, which is a bit confusing. It would be nice to have more consistency, though I can see to some extent why in some cases there is a named vector and in others a tidytable. I think it would be helpful for users to have a vignette with some more examples of working with the output to get it into more common shapes for next steps in an analysis, such as using `enframe` to turn a named vector into a tibble/tidytable if desired.

### Non-English Usage and Examples

As a linguistics-oriented package, it would be helpful to include some vignette examples that include languages other than English, and to mention any potential limitations for languages with different orthographic/tokenization issues. For example, Mandarin Chinese does not use spaces, but it might be possible to use the functions if spaces (or another delimiter) are indicated where word boundaries are desired. There are other issues that arise for the tokenization of these languages, however, as most (all?) of the models will be based on single character tokenization, which may not provide the best probabilities for words which are predominantly multi-character. These things are a little more clear if one is coding the Python for transformers, but are obscured by the wrapper functions, so seem relevant to at least point towards other resources on.

``` r
library(pangoling)
library(tidyverse)
```

### Specific Function Comments

#### `causal_lp`

The `causal_lp()` function extracts log probabilities from a language model, by default from GPT-2.

I find it confusing that the x argument can be used it 2 different ways, either as a full sentence where each word is a target, or as targets with a separate `l_contexts` prefix. The examples don’t make it clear how a non-atomic x vector would work with `l_contexts`:

``` r
causal_lp(
  x = "tree.",
  l_contexts = "The apple doesn't fall far from the",
  .by = NULL, # it's ignored anyways
  model = "gpt2"
)
```

    ## -1.581758

``` r
causal_lp(
  x = c("branch.", "tree."),
  l_contexts = "The apple doesn't fall far from the",
  .by = NULL, # it's ignored anyways
  model = "gpt2"
)
```

    ## -11.221713  -1.581758

``` r
causal_lp(
  x = c("branch.", "tree."),
  model = "gpt2"
)
```

    ##        NA -13.32788

The last one doesn’t make much sense, but it seems odd to me that it looks very similar to the one above it, yet what it’s actually checking is the probability of “tree.” following “branch.”.

I don’t really understand the `.by` documentation. The documentation seems to suggest it indicates the text separator, but I think it is only used as a grouping variable if one has multiple sentences in a dataframe.

The documentation is also not clear about what “batch size” is and how it should/could be used.

There is a typo in the ‘run examples’ where ‘tree.’ is given twice - `l-contexts` should just be “The apple doesn’t fall far from the” as follows:

``` r
output_causal_lp <- causal_lp(
  x = "tree.",
  l_contexts = "The apple doesn't fall far from the",
  .by = NULL, # it's ignored anyways
  model = "gpt2"
)
```

    ## `The apple doesn't fall far from the tree.`

Some ways to get long dataframe output from named vectors (recommended for documentation):

``` r
output_causal_lp <- causal_lp(
  x = c("The", "apple", "doesn't", "fall", "far", "from", "the", "tree."),
  model = "gpt2"
)
```

    ## `The apple doesn't fall far from the tree.`

``` r
tibble::enframe(output_causal_lp)
```

    ## 8 tree.    -1.58

``` r
data.frame(token = names(output_causal_lp), lp = output_causal_lp, row.names = NULL)
```

    ## 8   tree.  -1.5817576

#### `causal_lp_mats`

From what I can tell, this function outputs a matrix with probabilities of the full vocabulary at each word position in the `x` argument vector. This isn’t very clear from the help description though. I can’t tell from the example provided why this is output as a matrix rather than a dataframe.

The columns are based on the BPE tokens, not words (for example, there are separate columns for “doesn” and “’t”), so the title/description should not be about “possible word” but token instead. This should also be more clearly documented since the `causal_lp` output adds the relevant log probs to give the total word (I can see why this function is different, but it could be unexpected for the user).

It would be useful to have an example in the help or vignettes of how to run this one multiple sentences, and how to wrangle the output.

#### `causal_tokens_lp_tbl`

Returns for tokens instead of words.

It is confusing that the `texts` argument seems to be the same as the `x` argument in other functions - is there a reason for the difference? This comes back to my comment above about the `x` argument - perhaps it should be generally be separated into two arguments for the `causal_lp` functions, one which is like `texts` here and one which is just for `targets`. Why is list input an option here for `texts` but not for `x`? I don’t see an example of how it would be used.

I find the `.id` argument name to be unintuitive. For whatever reason `.id` seems to me like it is requiring an existing ID input, whereas it is asking for an output column name to be set.

#### `masked_lp`

Log probs using masked bidirectional models.

See my general comments above about preferring a different argument structure.

#### `masked_tokens_tbl`

This function returns the probabilities of all vocabulary tokens for the masked positions. Very useful, but a bit confusing how different this is from the similarly named `causal_tokens_lp_tbl` which gets probabilities for all of the tokens **in the sentence**. This one is more similar to `causal_lp_mats` but doesn’t return matrices (why the distinction?).

Maybe a naming distinction could be made for when a function is across the vocabulary vs. the “text” tokens?

### Test Failures

Failure output from `devtools::test()`:

    Failure (test-tr_causal.R:152:3): can handle extra parameters
    token_lp$token (`actual`) not equal to token_lp2$token[-1] (`expected`).

    `actual[1:4]`:   "This" "Ġisn" "'t" "Ġit"
    `expected[1:3]`:        "Ġisn" "'t" "Ġit"

    Failure (test-tr_causal.R:153:3): can handle extra parameters
    `token_lp2` (`actual`) not equal to `token_lp3` (`expected`).

      `attr(actual, 'row.names')[3:5]`: 3 4 5
    `attr(expected, 'row.names')[3:6]`: 3 4 5 6

    actual vs expected
                            token            lp
    - actual[1, ]   This                     NA
    - actual[2, ]   Ġisn          -6.7738981247
    - actual[3, ]   't            -0.0009036748
    - actual[4, ]   Ġit           -7.1741533279
    - actual[5, ]   .             -1.0767078400
    + expected[1, ] <|endoftext|>            NA
    + expected[2, ] This          -4.8580427170
    + expected[3, ] Ġisn          -5.5722875595
    + expected[4, ] 't            -0.0003263418
    + expected[5, ] Ġit           -6.7451515198
    + expected[6, ] .             -0.7511805296

    `actual$token[1:3]`:                   "This" "Ġisn" "'t"
    `expected$token[1:4]`: "<|endoftext|>" "This" "Ġisn" "'t"

        actual$lp             | expected$lp
    [1] NA                    | NA                    [1]
    [2] -6.77389812469482     - -4.85804271697998     [2]
    [3] -0.000903674808796495 - -5.57228755950928     [3]
    [4] -7.17415332794189     - -0.000326341774780303 [4]
    [5] -1.07670783996582     - -6.74515151977539     [5]
                              - -0.751180529594421    [6]

Also a warning:

    Warning (test-tr_causal.R:164:3): can handle extra parameters
    longer argument not a multiple of length of shorter
     1. pangoling::causal_lp(x = c("This", "is", "it"), add_special_tokens = TRUE)
          at test-tr_causal.R:164:2
     2. tidytable::pmap(...)
          at pangoling/R/tr_causal.R:217:2
     4. base::mapply(...)
     5. pangoling (local) `<fn>`(dots[[1L]][[1L]], dots[[2L]][[1L]], dots[[3L]][[1L]])
     6. pangoling:::word_lp(...)
          at pangoling/R/tr_causal.R:233:6
     7. tidytable::map2_dbl(...)
          at pangoling/R/tr_utils.R:324:2
     8. tidytable::map2(.x, .y, .f, ...)
     9. base::mapply(.f, .x, .y, MoreArgs = list(...), SIMPLIFY = FALSE)
