+++ .github/ISSUE_TEMPLATE/issue_template.md
@@ -3,7 +3,7 @@ name: Bug report or feature request
 about: Describe a bug you've seen or make a case for a new feature
 ---
 
-Please briefly describe your problem and what output you expect. If you have a question, please don't use this form. Instead, ask on <https://stackoverflow.com/> or <https://community.rstudio.com/>.
+Please briefly describe your problem and what output you expect. If you have a question, please don't use this form. Instead, ask on <https://github.com/bnicenboim/pangoling/discussions>.
 
 Please include a minimal reproducible example (AKA a reprex). If you've never heard of a [reprex](http://reprex.tidyverse.org/) before, start by reading <https://www.tidyverse.org/help/#reprex>.
 
