### 2023-06-06 25e3957
added contributing guidelines to the readme
