### 2023-06-06 a6a247b
helpful message when loads and sets cache
