+++ R/tr_utils.R
@@ -365,3 +365,42 @@ num_to_token <- function(x, tkzr) {
     tkzr$convert_ids_to_tokens(x)
   })
 }
+
+#' Set Cache Folder for HuggingFace Transformers
+#'
+#' This function sets the cache directory for HuggingFace transformers. If a path is given, the function checks if the directory exists and then sets the `TRANSFORMERS_CACHE` environment variable to this path.
+#' If no path is provided, the function checks for the existing cache directory in a number of environment variables.
+#' If none of these environment variables are set, it provides the user with information on the default cache directory.
+#'
+#' @param path Character string, the path to set as the cache directory. If NULL, the function will look for the cache directory in a number of environment variables. Default is NULL.
+#'
+#' @return Nothing is returned, this function is called for its side effect of setting the `TRANSFORMERS_CACHE` environment variable, or providing information to the user.
+#' @export
+#'
+#' @examples
+#' \dontrun{
+#' set_cache_folder("~/new_cache_dir")
+#' }
+#' @seealso \url{https://huggingface.co/docs/transformers/installation?highlight=transformers_cache#cache-setup}
+#' @references HuggingFace Transformers: \url{https://huggingface.co/transformers/index.html}
+set_cache_folder <- function(path = NULL){
+  if(!is.null(path)){
+    if(!dir.exists(path)) stop2("Folder '", path, "' doesn't exist.")
+    reticulate::py_run_string(paste0("import os\nos.environ['TRANSFORMERS_CACHE']='",path,"'"))
+    reticulate::py_run_string(paste0("import os\nos.environ['HF_HOME']='",path,"'"))
+}
+  path <- c(Sys.getenv("TRANSFORMERS_CACHE"),
+  Sys.getenv("HUGGINGFACE_HUB_CACHE"),
+  Sys.getenv("HF_HOME"),
+  Sys.getenv("XDG_CACHE_HOME"))
+
+  path <- paste0(path[path!=""],"")[1]
+  if(path != ""){
+    message_verbose("Pretrained models and tokenizers are downloaded and locally cached at '", path,"'.")
+    } else {
+      message_verbose("By default pretrained models are downloaded and locally cached at: ~/.cache/huggingface/hub. This is the default directory given by the shell environment variable TRANSFORMERS_CACHE. On Windows, the default directory is given by C:\\Users\\username\\.cache\\huggingface\\hub.
+
+For changing the shell environment variables that affect the cache folder see https://huggingface.co/docs/transformers/installation?highlight=transformers_cache#cache-setup")
+    }
+
+}
