+++ man/causal_preload.Rd
@@ -15,7 +15,7 @@ causal_preload(
 \arguments{
 \item{model}{Name of a pre-trained model or folder.}
 
-\item{checkpoint}{folder of a checkpoint.}
+\item{checkpoint}{Folder of a checkpoint.}
 
 \item{add_special_tokens}{Whether to include special tokens. It has the
 same default as the
