### 2023-06-07 #575 @bnicenboim
ok, I'm still struggling with docker to answer the main concern of @utkuturk and check how the installation goes. Hopefully, I'll be able to figure out how does an installation from scratch looks like.

In the meanwhile I have a question for @lisalevinson (but utku feel free to comment).

Lisa was confused with the two uses of `causal_lp`.

Ok, first load the packages.

``` r
library(pangoling)
library(tidytable)
```

There are two main formats that the causal_lp addresses. One is you have data word-by-word (or phrase-by-phrase) which is common stimuli for many self-paced reading, eye-tracking, ERP experiments that deal with naturalistic texts. (`other_word_level_stuff` represents some other info that one has about the words)

``` r
df_psychl <- tidytable(
  sent_n =
    c(1, 1, 1, 1, 1, 1, 1, 1, 2,2, 2, 2, 2, 2, 2),
  word = c("The", "apple", "doesn't","fall", "far", "from", "the", "tree.",
    "Don't", "judge", "a", "book", "by", "its", "cover."),    other_word_level_stuff = NA
)

df_psychl
#> # A tidytable: 15 × 3
#>    sent_n word    other_word_level_stuff
#>     <dbl> <chr>   <lgl>
#>  1      1 The     NA
#>  2      1 apple   NA
#>  3      1 doesn't NA
#>  4      1 fall    NA
#>  5      1 far     NA
#>  6      1 from    NA
#>  7      1 the     NA
#>  8      1 tree.   NA
#>  9      2 Don't   NA
#> 10      2 judge   NA
#> 11      2 a       NA
#> 12      2 book    NA
#> 13      2 by      NA
#> 14      2 its     NA
#> 15      2 cover.  NA
```
If the two sentences are not part of one single text, that's when you *must* use `.by`, otherwise it will think that the two sentences are related.

``` r
df_psychl <- df_psychl |>
  mutate(lp = causal_lp(x = word, .by = sent_n))
#> Processing using causal model ''...
#> Processing a batch of size 1 with 10 tokens.
#> Processing a batch of size 1 with 9 tokens.
#> Text id: 1
#> `The apple doesn't fall far from the tree.`
#> Text id: 2
#> `Don't judge a book by its cover.`
df_psychl
#> # A tidytable: 15 × 4
#>    sent_n word    other_word_level_stuff      lp
#>     <dbl> <chr>   <lgl>                    <dbl>
#>  1      1 The     NA                      NA
#>  2      1 apple   NA                     -10.9
#>  3      1 doesn't NA                      -5.50
#>  4      1 fall    NA                      -3.60
#>  5      1 far     NA                      -2.91
#>  6      1 from    NA                      -0.745
#>  7      1 the     NA                      -0.207
#>  8      1 tree.   NA                      -1.58
#>  9      2 Don't   NA                      NA
#> 10      2 judge   NA                      -6.27
#> 11      2 a       NA                      -2.33
#> 12      2 book    NA                      -1.97
#> 13      2 by      NA                      -0.409
#> 14      2 its     NA                      -0.257
#> 15      2 cover.  NA                      -1.38
```

The other use case is an experiment where the stimuli are also sentences, regardless of how you present the context,  you only care about the critical region:

```r
df_psychl2 <- tidytable(
  item_n =
    c(1, 2, 3, 4),
  context = c("The apple doesn't fall far from the",
              "The apple doesn't fall far from the",
              "Don't judge a book by its",
              "Don't judge a book by its"),
  critical = c("tree.","floor.","cover.","back."),
  other_word_level_stuff = NA
)

df_psychl2
#> # A tidytable: 4 × 4
#>   item_n context                             critical other_word_level_stuff
#>    <dbl> <chr>                               <chr>    <lgl>
#> 1      1 The apple doesn't fall far from the tree.    NA
#> 2      2 The apple doesn't fall far from the floor.   NA
#> 3      3 Don't judge a book by its           cover.   NA
#> 4      4 Don't judge a book by its           back.    NA
```

In that case each row is an item, and then `.by` is not needed, but you need to have the `l_context` argument.
```r
df_psychl2 <- df_psychl2 |>
  mutate(lp = causal_lp(x = critical, l_contexts = context))
#> Processing using causal model ''...
#> Ignoring `.by` argument
#> Processing a batch of size 1 with 10 tokens.
#> Processing a batch of size 1 with 10 tokens.
#> Processing a batch of size 1 with 9 tokens.
#> Processing a batch of size 1 with 9 tokens.
#> Text id: 1
#> `The apple doesn't fall far from the tree.`
#> Text id: 2
#> `The apple doesn't fall far from the floor.`
#> Text id: 3
#> `Don't judge a book by its cover.`
#> Text id: 4
#> `Don't judge a book by its back.`
df_psychl2
#> # A tidytable: 4 × 5
#>   item_n context                          critical other_word_level_stuff     lp
#>    <dbl> <chr>                            <chr>    <lgl>                   <dbl>
#> 1      1 The apple doesn't fall far from… tree.    NA                      -1.58
#> 2      2 The apple doesn't fall far from… floor.   NA                     -10.2
#> 3      3 Don't judge a book by its        cover.   NA                      -1.38
#> 4      4 Don't judge a book by its        back.    NA                     -17.2
```

Do you think that having better examples would help? Or that I should name the arguments differently or that I should have two different functions?

Also for both of you, do you think I should change `x`, to `targets` to match the masked models, or to change the `masked` models targets to `x`?

Also `.by` is the only argument with a dot, I have it like this to match [dplyr](https://dplyr.tidyverse.org/reference/summarise.html)/[tidytable](https://markfairbanks.github.io/tidytable/reference/summarize.html), should I keep it like it is? Or add dots everywhere or remove dots everywhere? I think tidyverse functions tend to have dots everywhere, so that would make sense, right?
