### 2023-06-13 #575 @bnicenboim
Regarding @utkuturk's comment about the installation.

I used docker to create fresh installation of R (using https://rocker-project.org/images/other/r-ubuntu.html) with
```
docker run -i -t rocker/r-ubuntu:focal
```
The only thing I had to do was to install `png` first for some reason pangoling couldn't install it, I'll check if it works if I add it explicitly as a dependency.  But other than that, everything worked automatically. I didn't have to install or configure python, or conda or anything. Everything was taken care automatically (as it should be), the package follows this (https://rstudio.github.io/reticulate/articles/python_dependencies.html#using-configreticulate) which in theory should work.

So basically, I had to do.
```
install.packages("png")
install.packages("remotes")
remotes::install_github("bnicenboim/pangoling")
```
And that's it.

Then the first time I run a command from pangoling, I was asked

```
Processing using causal model ''...
No non-system installation of Python could be found.
Would you like to download and install Miniconda?
Miniconda is an open source environment management system for Python.
See https://docs.conda.io/en/latest/miniconda.html for more details.

Would you like to install Miniconda? [Y/n]:
```
I answered `Y` and then it just worked.

@utkuturk, did you have any specific problem with the installation? Or were you worried that users without python and conda would not manage?
It's certainly possible that the automatic installation doesn't work with windows, is that the case? If so, I'll try to install a virtual machine to test that.
