### 2023-10-26 48a3b7d
troubleshooting vignette
