+++ man/perplexity_calc.Rd
@@ -34,4 +34,8 @@ perplexity_calc(probs, log.p = FALSE)
 lprobs <- log(probs)
 perplexity_calc(lprobs, log.p = TRUE)
 }
+\seealso{
+Other general functions: 
+\code{\link{set_cache_folder}()}
+}
 \concept{general functions}
