+++ README.md
@@ -16,7 +16,7 @@ public.](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostat
 Review](https://badges.ropensci.org/575_status.svg)](https://github.com/ropensci/software-review/issues/575)
 <!-- badges: end -->
 
-`pangoling`[^1] is an R package for estimating the log-probabilities of
+`pangoling`\[1\] is an R package for estimating the log-probabilities of
 words in a given context using transformer models. The package provides
 an interface for utilizing pre-trained transformer models (such as GPT-2
 or BERT) to obtain word probabilities. These log-probabilities are often
@@ -28,7 +28,7 @@ The package is mostly a wrapper of the python package
 [`transformers`](https://pypi.org/project/transformers/) to process data
 in a convenient format.
 
-## Important! Limitations and bias
+## Important\! Limitations and bias
 
 The training data of the most popular models (such as GPT-2) haven’t
 been released, so one cannot inspect it. It’s clear that the data
@@ -94,11 +94,11 @@ as follows:
 ``` r
 df_sent <- df_sent |>
   mutate(lp = causal_lp(word, .by = sent_n))
-#> Processing using causal model 'gpt2'...
-#> Processing 1 batch(es) of 10 tokens.
+#> Processing using causal model ''...
+#> Processing a batch of size 1 with 10 tokens.
+#> Processing a batch of size 1 with 9 tokens.
 #> Text id: 1
 #> `The apple doesn't fall far from the tree.`
-#> Processing 1 batch(es) of 9 tokens.
 #> Text id: 2
 #> `Don't judge a book by its cover.`
 df_sent
@@ -125,7 +125,7 @@ df_sent
 ## How to cite
 
 > Nicenboim B (2023). *pangoling: Access to language model predictions
-> in R*. R package version 0.0.0.9006, DOI:
+> in R*. R package version 0.0.0.9007, DOI:
 > [10.5281/zenodo.7637526](https://zenodo.org/badge/latestdoi/497831295),
 > <https://github.com/bnicenboim/pangoling>.
 
@@ -146,7 +146,7 @@ Another R package that act as a wrapper for
 [`text`](https://r-text.org//) However, `text` is more general, and its
 focus is on Natural Language Processing and Machine Learning.
 
-[^1]: The logo of the package was created with [stable
+1.  The logo of the package was created with [stable
     diffusion](https://huggingface.co/spaces/stabilityai/stable-diffusion)
     and the R package
     [hexSticker](https://github.com/GuangchuangYu/hexSticker).
