+++ pkgdown.yml
@@ -4,7 +4,8 @@ pkgdown_sha: ~
 articles:
   intro-bert: intro-bert.html
   intro-gpt2: intro-gpt2.html
-last_built: 2023-10-26T13:44Z
+  troubleshooting: troubleshooting.html
+last_built: 2023-10-26T17:05Z
 urls:
   reference: http://bruno.nicenboim.me/pangoling/reference
   article: http://bruno.nicenboim.me/pangoling/articles
