+++ .github/workflows/R-CMD-check.yaml
@@ -45,7 +45,7 @@ jobs:
           needs: check
       - uses: actions/setup-python@v4
         with:
-          python-version: "3.9"
+          python-version: "3.9.x"
 
       - name: setup r-reticulate venv
         shell: Rscript {0}
