+++ .github/workflows/test-coverage.yaml
@@ -28,7 +28,7 @@ jobs:
 
       - uses: actions/setup-python@v4
         with:
-          python-version: "3.x"
+          python-version: "3.9.x"
 
       - name: setup r-reticulate venv
         shell: Rscript {0}
