+++ tests/testthat/test-tr_causal.R
@@ -28,8 +28,8 @@ test_that("empty or small strings", {
   small_str <- c("It", "It", "is")
   lp_small <- causal_lp(x = small_str, .by = c(1, 2, 2))
   expect_equal(lp_small[1:2], c(It = NA_real_, It = NA_real_))
-  expect_warning(lp_small <- causal_lp(x = c("", "It"), .by = c(1, 2)))
-  expect_equal(lp_small, c(NA_real_, "It" = NA_real_))
+  expect_warning(lp_small_ <- causal_lp(x = c("", "It"), .by = c(1, 2)))
+  expect_equal(lp_small_, c(NA_real_, "It" = NA_real_))
 })
 
 if(0){
@@ -60,7 +60,9 @@ test_that("gpt2 get prob work", {
   expect_equal(cont[1]$token, "Ġtree")
   lp_prov <- causal_lp(x = prov_words)
   expect_equal(names(lp_prov), prov_words)
-  lp_cont <- causal_lp(l_contexts = c("Don't judge a book by its","The apple doesn't fall far from the"), x = c("cover", "tree"))
+  lp_cont <- causal_lp(l_contexts = c("Don't judge a book by its",
+                                      "The apple doesn't fall far from the"),
+                       x = c("cover", "tree"))
   expect_equal(lp_cont[2], lp_prov[8], tolerance = .0001)
   lp_sent2 <- causal_lp(x = sent2_words)
   expect_equal(names(lp_sent2), sent2_words)
@@ -110,6 +112,14 @@ test_that("gpt2 get prob work", {
     unname(lp_sent_rep[seq_along(sent_w)]),
     unname(lp_sent_rep[(length(sent_w) + 1):(2 * length(sent_w))])
   )
+ df_order1 <-  data.frame(word = c(sent2_words,prov_words),
+             item = c(rep(1, each = length(sent2_words)),
+                      rep(2, each= length(prov_words))))
+ df_order2 <-  data.frame(word = c(sent2_words,prov_words),
+                          item = c(rep(2, each = length(sent2_words)),
+                                   rep(1, each= length(prov_words))))
+ expect_equal(causal_lp(df_order1$word, .by = df_order1$item),
+  causal_lp(x = df_order2$word, .by = df_order2$item))
 })
 
 test_that("batches work", {
