+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   intro-bert: intro-bert.html
   intro-gpt2: intro-gpt2.html
   troubleshooting: troubleshooting.html
-last_built: 2023-10-26T20:31Z
+last_built: 2023-11-04T13:51Z
 urls:
   reference: http://bruno.nicenboim.me/pangoling/reference
   article: http://bruno.nicenboim.me/pangoling/articles
