+++ NAMESPACE
@@ -8,6 +8,7 @@ export(causal_lp_mats)
 export(causal_next_tokens_tbl)
 export(causal_preload)
 export(causal_tokens_lp_tbl)
+export(install_py_pangoling)
 export(masked_config)
 export(masked_lp)
 export(masked_preload)
