### 2024-03-06 #575 @bnicenboim
> Hi @bnicenboim, I'm checking in on issues that have been static for awhile. I see this issue was moving along alright and then fell silent.
>
> Let us know if there is progress to report or any questions you may have. We can move to to a "Hold" status, or keep it open as-is if you expect to continue the respond to the reviews.

Hi, I did a lot of progress (and unrelated changes) and I had some question for @lisalevinson that was never answered, but  regardless it's fine to have this on hold until my teaching is over and I can finalize the last things.
