+++ man/causal_next_tokens_tbl.Rd
@@ -78,8 +78,8 @@ causal_next_tokens_tbl(
 \seealso{
 Other causal model functions: 
 \code{\link{causal_config}()},
-\code{\link{causal_lp_mats}()},
 \code{\link{causal_lp}()},
+\code{\link{causal_lp_mats}()},
 \code{\link{causal_preload}()},
 \code{\link{causal_tokens_lp_tbl}()}
 }
