+++ DESCRIPTION
@@ -34,7 +34,7 @@ Suggests:
     covr,
     spelling
 Config/testthat/edition: 3
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.1
 Roxygen: list(markdown = TRUE)
 Depends: 
     R (>= 4.1.0)
