+++ R/utils.R
@@ -112,20 +112,22 @@ message_verbose <- function(...) {
   if (options()$pangoling.verbose > 0) message(...)
 }
 
+
+#' @noRd
 stop2 <- function(...) {
   stop(..., call. = FALSE)
 }
 
-#' #' Replacement of str_match
-#' #' @noRd
-#' chr_match <- function(string, pattern) {
-#'   matches <- regexec(pattern = pattern, text = string)
-#'   list_matches <- lapply(
-#'     regmatches(x = string, m = matches),
-#'     function(x) if (length(x) == 0) NA else x
-#'   )
-#'   do.call("rbind", list_matches)
-#' }
+# #' Replacement of str_match
+# #' @noRd
+# chr_match <- function(string, pattern) {
+#   matches <- regexec(pattern = pattern, text = string)
+#   list_matches <- lapply(
+#     regmatches(x = string, m = matches),
+#     function(x) if (length(x) == 0) NA else x
+#   )
+#   do.call("rbind", list_matches)
+# }
 
 
 #' Replacement of str_detect
@@ -135,9 +137,10 @@ chr_detect <- function(string, pattern, ignore.case = FALSE) {
 }
 
 
-#' #' @noRd
-#' message_debug <- function(...) {
-#'   if (options()$pangoling.verbose > 1) message(...)
-#' }
+# 
+# message_debug <- function(...) {
+#   if (options()$pangoling.verbose > 1) message(...)
+# }
 
+#' @noRd
 "%||%" <- function(x, y) if (is.null(x)) y else x
