+++ DESCRIPTION
@@ -32,7 +32,8 @@ Suggests:
     testthat (>= 3.0.0),
     tictoc,
     covr,
-    spelling
+    spelling,
+    rstudioapi
 Config/testthat/edition: 3
 RoxygenNote: 7.3.1
 Roxygen: list(markdown = TRUE)
