### 2024-11-10 #575 @emilyriederer
Hi @bnicenboim - checking in with the rOpenSci editors team. Do you still plan on picking this back up?
