+++ R/zzz.R
@@ -43,6 +43,5 @@ torch <- NULL
     packageStartupMessage(pkgname, 
     " version ", 
     utils::packageVersion(pkgname),
-    "\nAn introduction to the package can be found in https://bruno.nicenboim.me/pangoling/articles/\n Notice that pretrained models and tokenizers are downloaded from https://huggingface.co/ the first time they are used.\n For changing the cache folder use:\n
-set_cache_folder(my_new_path)")
+    "\nAn introduction to the package can be found in https://bruno.nicenboim.me/pangoling/articles/\n Notice that pretrained models and tokenizers are downloaded from https://huggingface.co/ the first time they are used.\n For changing the cache folder use:\n set_cache_folder(my_new_path)")
 }
