### 2024-11-11 #575 @bnicenboim
Yes, I'm picking it up. Sorry for the long delay. I also wanted to add some other features. I'll see how easy it is or if at least the function names that I'll have are compatible with these other usages I'm thinking about.

 Also, I don't think my answer to @lisalevinson  is relevant anymore. She's completely right in that the two uses of causal_lp are confusing. I'll divide it into two functions.
