### 2024-11-11 22767bd
deprecated function, casual_each_pred created
