+++ man/causal_config.Rd
@@ -55,6 +55,7 @@ causal_config(model = "gpt2")
 }
 \seealso{
 Other causal model functions: 
+\code{\link{causal_each_pred}()},
 \code{\link{causal_lp}()},
 \code{\link{causal_lp_mats}()},
 \code{\link{causal_next_tokens_tbl}()},
