+++ man/causal_preload.Rd
@@ -64,6 +64,7 @@ causal_preload(model = "gpt2")
 \seealso{
 Other causal model functions: 
 \code{\link{causal_config}()},
+\code{\link{causal_each_pred}()},
 \code{\link{causal_lp}()},
 \code{\link{causal_lp_mats}()},
 \code{\link{causal_next_tokens_tbl}()},
