+++ man/causal_each_pred.Rd
@@ -1,13 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/tr_causal.R
-\name{causal_lp_mats}
-\alias{causal_lp_mats}
-\title{Get a list of matrices with the log probabilities of possible word given its previous context using a causal transformer}
+\name{causal_each_pred}
+\alias{causal_each_pred}
+\title{Get the predictability of each element of a vector of words (or phrases) using a causal transformer}
 \usage{
-causal_lp_mats(
+causal_each_pred(
   x,
   by = rep(1, length(x)),
-  sorted = FALSE,
+  log.p = getOption("pangoling.causal.log.p"),
+  ignore_regex = "",
   model = getOption("pangoling.causal.default"),
   checkpoint = NULL,
   add_special_tokens = NULL,
@@ -22,7 +23,14 @@ causal_lp_mats(
 
 \item{by}{Vector that indicates how the text should be split.}
 
-\item{sorted}{When default FALSE it will retain the order of groups we are splitting on. When TRUE then sorted (according to \code{by}) list(s) are returned.}
+\item{log.p}{If TRUE (default),  x are assumed to be log-transformed
+probabilities with base e, if FALSE x are assumed to be
+raw probabilities, alternatively log.p can be the base of
+other logarithmic transformations.}
+
+\item{ignore_regex}{Can ignore certain characters when calculates the log
+probabilities. For example \verb{^[[:punct:]]$} will ignore
+all punctuation  that stands alone in a token.}
 
 \item{model}{Name of a pre-trained model or folder.}
 
@@ -43,11 +51,12 @@ processing but take more memory.}
 \item{...}{not in use.}
 }
 \value{
-A list of matrices with tokens in their columns and the vocabulary of the model in their rows
+A named vector of log probabilities.
 }
 \description{
-Get a list of matrices with the log probabilities of possible word given
-its previous context using a causal transformer model.
+Get the predictability (default natural logarithm of the word probability) of each element of a vector of words (or phrases) using a causal transformer model. See the
+\href{https://bruno.nicenboim.me/pangoling/articles/intro-gpt2.html}{online article}
+in pangoling website for more examples.
 }
 \details{
 A causal language model (also called GPT-like, auto-regressive, or decoder
@@ -82,7 +91,7 @@ in pangoling website for more examples.
 
 \examples{
 \dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
-causal_lp_mats(
+causal_lp(
   x = c("The", "apple", "doesn't", "fall", "far", "from", "the", "tree."),
   model = "gpt2"
 )
@@ -91,8 +100,8 @@ causal_lp_mats(
 \seealso{
 Other causal model functions: 
 \code{\link{causal_config}()},
-\code{\link{causal_each_pred}()},
 \code{\link{causal_lp}()},
+\code{\link{causal_lp_mats}()},
 \code{\link{causal_next_tokens_tbl}()},
 \code{\link{causal_preload}()},
 \code{\link{causal_tokens_lp_tbl}()}
