+++ man/causal_tokens_lp_tbl.Rd
@@ -84,6 +84,7 @@ causal_tokens_lp_tbl(
 \seealso{
 Other causal model functions: 
 \code{\link{causal_config}()},
+\code{\link{causal_each_pred}()},
 \code{\link{causal_lp}()},
 \code{\link{causal_lp_mats}()},
 \code{\link{causal_next_tokens_tbl}()},
