+++ man/causal_lp.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/tr_causal.R
+% Please edit documentation in R/deprecated.R
 \name{causal_lp}
 \alias{causal_lp}
 \title{Get the log probability of each element of a vector of words (or phrases) using a causal transformer}
@@ -105,6 +105,7 @@ causal_lp(
 \seealso{
 Other causal model functions: 
 \code{\link{causal_config}()},
+\code{\link{causal_each_pred}()},
 \code{\link{causal_lp_mats}()},
 \code{\link{causal_next_tokens_tbl}()},
 \code{\link{causal_preload}()},
