### 2024-11-11 bfafb73
masked also deals with pred and not logp
