+++ man/causal_tokens_pred_tbl.Rd
@@ -1,11 +1,12 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/tr_causal.R
-\name{causal_tokens_lp_tbl}
-\alias{causal_tokens_lp_tbl}
-\title{Get the log probability of each token in a sentence (or group of sentences) using a causal transformer}
+\name{causal_tokens_pred_tbl}
+\alias{causal_tokens_pred_tbl}
+\title{Get the predictability of each token in a sentence (or group of sentences) using a causal transformer}
 \usage{
-causal_tokens_lp_tbl(
+causal_tokens_pred_tbl(
   texts,
+  log.p = getOption("pangoling.log.p"),
   model = getOption("pangoling.causal.default"),
   checkpoint = NULL,
   add_special_tokens = NULL,
@@ -31,16 +32,13 @@ model from Hugging Face is accessed.}
 
 \item{config_tokenizer}{List with other arguments that control how the tokenizer from Hugging Face is accessed.}
 
-\item{batch_size}{Maximum size of the batch. Larges batches speedup
-processing but take more memory.}
-
 \item{.id}{Name of the column with the sentence id.}
 }
 \value{
 A table with token names (\code{token}), log-probability (\code{lp}) and optionally sentence id.
 }
 \description{
-Get the log probability of each token in a sentence (or group of sentences) using a causal transformer model.
+Get the predictability of each token in a sentence (or group of sentences) using a causal transformer model.
 }
 \details{
 A causal language model (also called GPT-like, auto-regressive, or decoder
@@ -75,7 +73,7 @@ in pangoling website for more examples.
 
 \examples{
 \dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
-causal_tokens_lp_tbl(
+causal_tokens_pred_tbl(
   texts = c("The apple doesn't fall far from the tree."),
   model = "gpt2"
 )
@@ -84,10 +82,10 @@ causal_tokens_lp_tbl(
 \seealso{
 Other causal model functions: 
 \code{\link{causal_config}()},
-\code{\link{causal_each_pred}()},
 \code{\link{causal_lp}()},
-\code{\link{causal_lp_mats}()},
 \code{\link{causal_next_tokens_tbl}()},
-\code{\link{causal_preload}()}
+\code{\link{causal_pred_mats}()},
+\code{\link{causal_preload}()},
+\code{\link{causal_words_pred}()}
 }
 \concept{causal model functions}
