### 2024-11-11 3e9f30d
document options, pred rather than log.p is used now
