+++ man/masked_config.Rd
@@ -49,8 +49,8 @@ masked_config(model = "bert-base-uncased")
 }
 \seealso{
 Other masked model functions: 
-\code{\link{masked_lp}()},
 \code{\link{masked_preload}()},
+\code{\link{masked_targets_pred}()},
 \code{\link{masked_tokens_tbl}()}
 }
 \concept{masked model functions}
