+++ man/masked_tokens_tbl.Rd
@@ -72,7 +72,7 @@ masked_tokens_tbl("The [MASK] doesn't fall far from the tree.",
 \seealso{
 Other masked model functions: 
 \code{\link{masked_config}()},
-\code{\link{masked_lp}()},
-\code{\link{masked_preload}()}
+\code{\link{masked_preload}()},
+\code{\link{masked_targets_pred}()}
 }
 \concept{masked model functions}
