+++ man/masked_targets_pred.Rd
@@ -1,13 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/tr_masked.R
-\name{masked_lp}
-\alias{masked_lp}
+\name{masked_targets_pred}
+\alias{masked_targets_pred}
 \title{Get the log probability of a target word (or phrase) given a left and right context}
 \usage{
-masked_lp(
+masked_targets_pred(
   l_contexts,
   targets,
   r_contexts,
+  log.p = getOption("pangoling.log.p"),
   ignore_regex = "",
   model = getOption("pangoling.masked.default"),
   add_special_tokens = NULL,
@@ -21,6 +22,11 @@ masked_lp(
 \item{targets}{Target words.}
 
 \item{r_contexts}{Right context of the target word.}
+
+\item{log.p}{If TRUE (default),  x are assumed to be log-transformed
+probabilities with base e, if FALSE x are assumed to be
+raw probabilities, alternatively log.p can be the base of
+other logarithmic transformations.}
 
 \item{ignore_regex}{Can ignore certain characters when calculates the log
 probabilities. For example \verb{^[[:punct:]]$} will ignore
@@ -74,7 +80,7 @@ in pangoling website for more examples.
 
 \examples{
 \dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
-masked_lp(
+masked_targets_pred(
   l_contexts = c("The", "The"),
   targets = c("apple", "pear"),
   r_contexts = c(
