+++ NAMESPACE
@@ -11,8 +11,8 @@ export(causal_tokens_pred_tbl)
 export(causal_words_pred)
 export(install_py_pangoling)
 export(masked_config)
-export(masked_lp)
 export(masked_preload)
+export(masked_targets_pred)
 export(masked_tokens_tbl)
 export(ntokens)
 export(perplexity_calc)
