+++ R/tr_causal.R
@@ -104,19 +104,19 @@ causal_next_tokens_tbl <- function(context,
   message_verbose("Processing using causal model '", file.path(model, checkpoint), "'...")
   trf <- lang_model(model,
                     checkpoint = checkpoint,
-    task = "causal",
-    config_model = config_model
-  )
+                    task = "causal",
+                    config_model = config_model
+                    )
   tkzr <- tokenizer(model,
-    add_special_tokens = add_special_tokens,
-    config_tokenizer = config_tokenizer
-  )
+                    add_special_tokens = add_special_tokens,
+                    config_tokenizer = config_tokenizer
+                    )
 
   # no batches allowed
   context_tensor <- encode(list(unlist(context)),
-    tkzr,
-    add_special_tokens = add_special_tokens
-  )$input_ids
+                           tkzr,
+                           add_special_tokens = add_special_tokens
+                           )$input_ids
   generated_outputs <- trf(context_tensor)
   n_tokens <- length(context_tensor$tolist()[0])
   logits_next_word <- generated_outputs$logits[0][n_tokens - 1]
@@ -131,7 +131,7 @@ causal_next_tokens_tbl <- function(context,
 
 #' Get the predictability of each element of a vector of words (or phrases) using a causal transformer
 #'
-#' Get the predictability (default natural logarithm of the word probability) of each element of a vector of words (or phrases) using a causal transformer model. See the
+#' Get the predictability (by default the natural logarithm of the word probability) of each element of a vector of words (or phrases) using a causal transformer model. See the
 #' [online article](https://bruno.nicenboim.me/pangoling/articles/intro-gpt2.html)
 #' in pangoling website for more examples.
 #'
@@ -167,16 +167,16 @@ causal_next_tokens_tbl <- function(context,
 #' @family causal model functions
 #' @export
 causal_words_pred <- function(x,
-                      by = rep(1, length(x)),
-                      log.p = getOption("pangoling.log.p"),
-                      ignore_regex = "",
-                      model = getOption("pangoling.causal.default"),
-                      checkpoint = NULL,
-                      add_special_tokens = NULL,
-                      config_model = NULL,
-                      config_tokenizer = NULL,
-                      batch_size = 1,
-                      ...) {
+                              by = rep(1, length(x)),
+                              log.p = getOption("pangoling.log.p"),
+                              ignore_regex = "",
+                              model = getOption("pangoling.causal.default"),
+                              checkpoint = NULL,
+                              add_special_tokens = NULL,
+                              config_model = NULL,
+                              config_tokenizer = NULL,
+                              batch_size = 1,
+                              ...) {
   dots <- list(...)
   # Check for unknown arguments
   if (length(dots) > 0) {
@@ -196,14 +196,14 @@ causal_words_pred <- function(x,
     function(word) paste0(word, collapse = " ")
   )
   tkzr <- tokenizer(model,
-    add_special_tokens = add_special_tokens,
-    config_tokenizer = config_tokenizer
-  )
+                    add_special_tokens = add_special_tokens,
+                    config_tokenizer = config_tokenizer
+                    )
   trf <- lang_model(model,
                     checkpoint = checkpoint,
                     task = "causal",
                     config_model = config_model
-  )
+                    )
   tensors <- create_tensor_lst(
     texts = unname(pasted_texts),
     tkzr = tkzr,
@@ -214,51 +214,51 @@ causal_words_pred <- function(x,
 
   lmats <- lapply(tensors, function(tensor) {
     causal_mat(tensor,
-      trf,
-      tkzr,
-      add_special_tokens = add_special_tokens,
-      stride = stride
-    )
[…]
