+++ tests/testthat/test-tr_causal.R
@@ -21,14 +21,14 @@ test_that("gpt2 load and gets config", {
 
 test_that("empty or small strings", {
   skip_if_no_python_stuff()
-  lp_it <- causal_tokens_lp_tbl(texts = "It")
+  lp_it <- causal_tokens_pred_tbl(texts = "It")
   expect_equal(as.data.frame(lp_it), data.frame(token = "It", lp = NA_real_))
-  expect_warning(lp_NA <- causal_tokens_lp_tbl(texts = ""))
+  expect_warning(lp_NA <- causal_tokens_pred_tbl(texts = ""))
   expect_equal(as.data.frame(lp_NA), data.frame(token = "", lp = NA_real_))
   small_str <- c("It", "It", "is")
-  lp_small <- causal_lp(x = small_str, by = c(1, 2, 2))
+  lp_small <- causal_words_pred(x = small_str, by = c(1, 2, 2))
   expect_equal(lp_small[1:2], c(It = NA_real_, It = NA_real_))
-  expect_warning(lp_small_ <- causal_lp(x = c("", "It"), by = c(1, 2)))
+  expect_warning(lp_small_ <- causal_words_pred(x = c("", "It"), by = c(1, 2)))
   expect_equal(lp_small_, c(NA_real_, "It" = NA_real_))
 })
 
@@ -39,17 +39,17 @@ test_that("long input work", {
   long0 <- paste(rep("x", 1022), collapse = " ")
   long <- paste(rep("x", 1024), collapse = " ")
   longer <- paste(rep("x", 1025), collapse = " ")
-  lp_long0 <- causal_tokens_lp_tbl(texts = c(long0, long, longer), add_special_tokens = TRUE, batch_size = 3, model = "sshleifer/tiny-gpt2")
+  lp_long0 <- causal_tokens_pred_tbl(texts = c(long0, long, longer), add_special_tokens = TRUE, batch_size = 3, model = "sshleifer/tiny-gpt2")
   skip_on_os("windows") #the following just doesn't work on windows,
   # but it's not that important
-    lp_long1 <- causal_tokens_lp_tbl(c(long0, long, longer), add_special_tokens = TRUE, batch_size = 1, model = "sshleifer/tiny-gpt2")
+    lp_long1 <- causal_tokens_pred_tbl(c(long0, long, longer), add_special_tokens = TRUE, batch_size = 1, model = "sshleifer/tiny-gpt2")
   expect_equal(lp_long0, lp_long1)
 })
 }
 
 test_that("errors work", {
   skip_if_no_python_stuff()
-  expect_error(causal_lp(c("It", "is."), by = 3))
+  expect_error(causal_words_pred(c("It", "is."), by = 3))
 })
 
 test_that("gpt2 get prob work", {
@@ -58,18 +58,18 @@ test_that("gpt2 get prob work", {
     causal_next_tokens_tbl(context = "The apple doesn't fall far from the")
   expect_equal(sum(exp(cont$lp)), 1, tolerance = .0001)
   expect_equal(cont[1]$token, "Ġtree")
-  lp_prov <- causal_lp(x = prov_words)
+  lp_prov <- causal_words_pred(x = prov_words)
   expect_equal(names(lp_prov), prov_words)
-  lp_cont <- causal_lp(l_contexts = c("Don't judge a book by its",
+  lp_cont <- causal_targets_pred(l_contexts = c("Don't judge a book by its",
                                       "The apple doesn't fall far from the"),
-                       x = c("cover", "tree"))
+                       targets = c("cover", "tree"))
[…]
