+++ man/causal_lp.Rd
@@ -108,6 +108,7 @@ Other causal model functions:
 \code{\link{causal_next_tokens_tbl}()},
 \code{\link{causal_pred_mats}()},
 \code{\link{causal_preload}()},
+\code{\link{causal_targets_pred}()},
 \code{\link{causal_tokens_pred_tbl}()},
 \code{\link{causal_words_pred}()}
 }
