+++ R/tr_causal.R
@@ -47,8 +47,8 @@ causal_preload <- function(model = getOption("pangoling.causal.default"),
                            add_special_tokens = NULL,
                            config_model = NULL, config_tokenizer = NULL) {
   message_verbose("Preloading causal model ", model, "...")
-  lang_model(model, checkpoint = checkpoint, task = "causal", config_model)
-  tokenizer(model, add_special_tokens = add_special_tokens, config_tokenizer)
+  lang_model(model, checkpoint = checkpoint, task = "causal", config_model = config_model)
+  tokenizer(model, add_special_tokens = add_special_tokens, config_tokenizer = config_tokenizer)
   invisible()
 }
 
@@ -87,19 +87,20 @@ causal_config <- function(model = getOption("pangoling.causal.default"),
 #' @inherit  causal_preload details
 #' @return A table with possible next tokens and their log-probabilities.
 #' @examplesIf interactive()
-#' causal_next_tokens_tbl(
+#' causal_next_tokens_pred_tbl(
 #'   context = "The apple doesn't fall far from the",
 #'   model = "gpt2"
 #' )
 #'
 #' @family causal model functions
 #' @export
-causal_next_tokens_tbl <- function(context,
-                                   model = getOption("pangoling.causal.default"),
-                                   checkpoint = NULL,
-                                   add_special_tokens = NULL,
-                                   config_model = NULL,
-                                   config_tokenizer = NULL) {
+causal_next_tokens_pred_tbl <- function(context,
+                                        log.p = getOption("pangoling.log.p"),
+                                        model = getOption("pangoling.causal.default"),
+                                        checkpoint = NULL,
+                                        add_special_tokens = NULL,
+                                        config_model = NULL,
+                                        config_tokenizer = NULL) {
   if (length(unlist(context)) > 1) stop2("Only one context is allowed in this function.")
   message_verbose("Processing using causal model '", file.path(model, checkpoint), "'...")
   trf <- lang_model(model,
@@ -124,8 +125,9 @@ causal_next_tokens_tbl <- function(context,
   lp <- reticulate::py_to_r(l_softmax) |>
     unlist()
   vocab <- get_vocab(tkzr)
-  tidytable::tidytable(token = vocab, lp = lp) |>
-    tidytable::arrange(-lp)
+  tidytable::tidytable(token = vocab,
+                       pred = lp |> ln_p_change(log.p = log.p)) |>
+    tidytable::arrange(-pred)
 }
 
 
@@ -138,11 +140,12 @@ causal_next_tokens_tbl <- function(context,
 #'
 #' @param x Vector of words, phrases or texts.
 #' @param by Vector that indicates how the text should be split.
-#' @param log.p If TRUE (default),  x are assumed to be log-transformed
-#'                probabilities with base e, if FALSE x are assumed to be
-#'                raw probabilities, alternatively log.p can be the base of
-#'                other logarithmic transformations.
-#' @param ... not in use.        
+#' @param log.p Base of the logarithm used for the output predictability values.
+#'              If `TRUE` (default), the natural logarithm (base *e*) is used.
+#'              If `FALSE`, the raw probabilities are returned.
+#'              Alternatively, `log.p` can be set to a numeric value specifying
+#'              the base of the logarithm (e.g., `2` for base-2 logarithms).
+#' @param ... not in use.
 #' @inheritParams causal_preload
 #' @param ignore_regex Can ignore certain characters when calculates the log
[…]
