+++ man/causal_config.Rd
@@ -56,7 +56,7 @@ causal_config(model = "gpt2")
 \seealso{
 Other causal model functions: 
 \code{\link{causal_lp}()},
-\code{\link{causal_next_tokens_tbl}()},
+\code{\link{causal_next_tokens_pred_tbl}()},
 \code{\link{causal_pred_mats}()},
 \code{\link{causal_preload}()},
 \code{\link{causal_targets_pred}()},
