+++ vignettes/articles/intro-gpt2.Rmd
@@ -28,13 +28,13 @@ library(tidytable) # fast alternative to dplyr
 library(tictoc) # measure time
 ```
 
-Then let's examine which continuation GPT-2 predicts following a specific context. [Hugging Face](https://huggingface.co/) provide access to pre-trained models, including freely available versions of different sizes of  [GPT-2](https://huggingface.co/gpt2). The function `causal_next_tokens_tbl()` will, by default, use the smallest version of [GPT-2](https://huggingface.co/gpt2), but this can be modified with the argument `model`.
+Then let's examine which continuation GPT-2 predicts following a specific context. [Hugging Face](https://huggingface.co/) provide access to pre-trained models, including freely available versions of different sizes of  [GPT-2](https://huggingface.co/gpt2). The function `causal_next_tokens_pred_tbl()` will, by default, use the smallest version of [GPT-2](https://huggingface.co/gpt2), but this can be modified with the argument `model`.
 
 Let's see what GPT-2 predicts following "The apple doesn't fall far from the".
 
 ```{r}
 tic()
-(df_pred <- causal_next_tokens_tbl("The apple doesn't fall far from the"))
+(df_pred <- causal_next_tokens_pred_tbl("The apple doesn't fall far from the"))
 toc()
 ```
 
@@ -45,7 +45,7 @@ The first time a model is run, it will download some files that will be availabl
 
 ```{r}
 tic()
-(df_pred <- causal_next_tokens_tbl("The apple doesn't fall far from the"))
+(df_pred <- causal_next_tokens_pred_tbl("The apple doesn't fall far from the"))
 toc()
 ```
 
@@ -64,7 +64,7 @@ tokenize_lst("This is different from This")
 ```
 
 
-Going back to the initial example, because `causal_next_tokens_tbl` returns log (natural) probabilities, if we exponentiate them and we sum them, we should get 1:
+Going back to the initial example, because `causal_next_tokens_pred_tbl` returns log (natural) probabilities, if we exponentiate them and we sum them, we should get 1:
 
 ```{r}
 sum(exp(df_pred$lp))
@@ -78,7 +78,7 @@ When doing tests, `sshleifer/tiny-gpt2` is quite useful since it's tiny. But not
 ```{r}
 causal_preload("sshleifer/tiny-gpt2")
 tic()
-causal_next_tokens_tbl("The apple doesn't fall far from the",
+causal_next_tokens_pred_tbl("The apple doesn't fall far from the",
   model = "sshleifer/tiny-gpt2"
 )
 toc()
