+++ R/deprecated.R
@@ -1,3 +1,242 @@
+#' Get the possible tokens and their log probabilities for each mask in a sentence
+#'
+#' For each mask in a sentence, get the possible tokens and their log
+#' probabilities using a masked transformer.
+#'
+#' @section More examples:
+#' See the
+#' [online article](https://bruno.nicenboim.me/pangoling/articles/intro-bert.html)
+#' in pangoling website for more examples.
+#'
+#'
+#' @param masked_sentences Masked sentences.
+#' @inheritParams masked_preload
+#' @inherit masked_preload details
+#' @return A table with the masked sentences, the tokens (`token`),
+#'         log probability (`lp`), and the respective mask number (`mask_n`).
+#' @examplesIf interactive()
+#' masked_tokens_tbl("The [MASK] doesn't fall far from the tree.",
+#'   model = "bert-base-uncased"
+#' )
+#'
+#' @family masked model functions
+#' @export
+masked_tokens_tbl <- function(masked_sentences,
+                              model = getOption("pangoling.masked.default"),
+                              add_special_tokens = NULL,
+                              config_model = NULL,
+                              config_tokenizer = NULL) {
+  message_verbose("Processing using masked model '", model, "'...")
+  tkzr <- tokenizer(model,
+    add_special_tokens = add_special_tokens,
+    config_tokenizer = config_tokenizer
+  )
+  trf <- lang_model(model,
+    task = "masked",
+    config_model = config_model
+  )
+  vocab <- get_vocab(tkzr)
+  # non_batched:
+  # TODO: speedup using batches
+  tidytable::map_dfr(masked_sentences, function(masked_sentence) {
+    masked_tensor <- encode(list(masked_sentence), tkzr,
+      add_special_tokens = add_special_tokens
+    )$input_ids
+    outputs <- trf(masked_tensor)
+    mask_pos <- which(masked_tensor$tolist()[[1]] == tkzr$mask_token_id)
+    logits_masks <- outputs$logits[0][mask_pos - 1] # python starts in 0
+    lp <- reticulate::py_to_r(
+      torch$log_softmax(logits_masks, dim = -1L)$tolist()
+    )
+    if (length(mask_pos) <= 1) lp <- list(lp) # to keep it consistent
+    # names(lp) <-  1:length(lp)
+    if (length(mask_pos) == 0) {
+      tidytable::tidytable(
+        masked_sentence = masked_sentence,
+        token = NA,
+        lp = NA,
+        mask_n = NA
+      )
+    } else {
+      lp |> tidytable::map_dfr(~
+        tidytable::tidytable(
+          masked_sentence = masked_sentence,
+          token = vocab, lp = .x
+        ) |>
+          tidytable::arrange(-lp), .id = "mask_n")
+    }
+  }) |>
+    tidytable::relocate(mask_n, .after = tidyselect::everything())
+}
+
+#' Get the log probability of a target word (or phrase) given a left and right context
+#'
+#' Get the log probability of a vector of target words (or phrase) given a
+#' vector of left and of right contexts using a masked transformer.
+#'
+#' @section More examples:
+#' See the
+#' [online article](https://bruno.nicenboim.me/pangoling/articles/intro-bert.html)
+#' in pangoling website for more examples.
+#'
+#'
+#' @param l_contexts Left context of the target word.
+#' @param targets Target words.
+#' @param r_contexts Right context of the target word.
+#' @inheritParams masked_preload
+#' @inheritParams causal_lp
+#' @inherit masked_preload details
+#' @return A named vector of log probabilities.
+#' @examplesIf interactive()
+#' masked_lp(
+#'   l_contexts = c("The", "The"),
+#'   targets = c("apple", "pear"),
+#'   r_contexts = c(
[…]
