+++ man/masked_tokens_tbl.Rd
@@ -1,9 +1,17 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/tr_masked.R
+% Please edit documentation in R/deprecated.R, R/tr_masked.R
 \name{masked_tokens_tbl}
 \alias{masked_tokens_tbl}
 \title{Get the possible tokens and their log probabilities for each mask in a sentence}
 \usage{
+masked_tokens_tbl(
+  masked_sentences,
+  model = getOption("pangoling.masked.default"),
+  add_special_tokens = NULL,
+  config_model = NULL,
+  config_tokenizer = NULL
+)
+
 masked_tokens_tbl(
   masked_sentences,
   model = getOption("pangoling.masked.default"),
@@ -29,8 +37,14 @@ model from Hugging Face is accessed.}
 \value{
 A table with the masked sentences, the tokens (\code{token}),
 log probability (\code{lp}), and the respective mask number (\code{mask_n}).
+
+A table with the masked sentences, the tokens (\code{token}),
+log probability (\code{lp}), and the respective mask number (\code{mask_n}).
 }
 \description{
+For each mask in a sentence, get the possible tokens and their log
+probabilities using a masked transformer.
+
 For each mask in a sentence, get the possible tokens and their log
 probabilities using a masked transformer.
 }
@@ -60,9 +74,19 @@ check the status of
 See the
 \href{https://bruno.nicenboim.me/pangoling/articles/intro-bert.html}{online article}
 in pangoling website for more examples.
+
+
+See the
+\href{https://bruno.nicenboim.me/pangoling/articles/intro-bert.html}{online article}
+in pangoling website for more examples.
 }
 
 \examples{
+\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+masked_tokens_tbl("The [MASK] doesn't fall far from the tree.",
+  model = "bert-base-uncased"
+)
+\dontshow{\}) # examplesIf}
 \dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 masked_tokens_tbl("The [MASK] doesn't fall far from the tree.",
   model = "bert-base-uncased"
@@ -72,6 +96,13 @@ masked_tokens_tbl("The [MASK] doesn't fall far from the tree.",
 \seealso{
 Other masked model functions: 
 \code{\link{masked_config}()},
+\code{\link{masked_lp}()},
+\code{\link{masked_preload}()},
+\code{\link{masked_targets_pred}()}
+
+Other masked model functions: 
+\code{\link{masked_config}()},
+\code{\link{masked_lp}()},
 \code{\link{masked_preload}()},
 \code{\link{masked_targets_pred}()}
 }
