+++ man/masked_targets_pred.Rd
@@ -95,6 +95,7 @@ masked_targets_pred(
 \seealso{
 Other masked model functions: 
 \code{\link{masked_config}()},
+\code{\link{masked_lp}()},
 \code{\link{masked_preload}()},
 \code{\link{masked_tokens_tbl}()}
 }
