+++ man/causal_words_pred.Rd
@@ -107,10 +107,13 @@ causal_words_pred(
 Other causal model functions: 
 \code{\link{causal_config}()},
 \code{\link{causal_lp}()},
+\code{\link{causal_lp_mats}()},
 \code{\link{causal_next_tokens_pred_tbl}()},
+\code{\link{causal_next_tokens_tbl}()},
 \code{\link{causal_pred_mats}()},
 \code{\link{causal_preload}()},
 \code{\link{causal_targets_pred}()},
+\code{\link{causal_tokens_lp_tbl}()},
 \code{\link{causal_tokens_pred_tbl}()}
 }
 \concept{causal model functions}
