+++ man/causal_tokens_pred_tbl.Rd
@@ -92,10 +92,13 @@ causal_tokens_pred_tbl(
 Other causal model functions: 
 \code{\link{causal_config}()},
 \code{\link{causal_lp}()},
+\code{\link{causal_lp_mats}()},
 \code{\link{causal_next_tokens_pred_tbl}()},
+\code{\link{causal_next_tokens_tbl}()},
 \code{\link{causal_pred_mats}()},
 \code{\link{causal_preload}()},
 \code{\link{causal_targets_pred}()},
+\code{\link{causal_tokens_lp_tbl}()},
 \code{\link{causal_words_pred}()}
 }
 \concept{causal model functions}
