+++ NAMESPACE
@@ -4,14 +4,18 @@ S3method(tokenize_lst,character)
 S3method(tokenize_lst,numeric)
 export(causal_config)
 export(causal_lp)
+export(causal_lp_mats)
 export(causal_next_tokens_pred_tbl)
+export(causal_next_tokens_tbl)
 export(causal_pred_mats)
 export(causal_preload)
 export(causal_targets_pred)
+export(causal_tokens_lp_tbl)
 export(causal_tokens_pred_tbl)
 export(causal_words_pred)
 export(install_py_pangoling)
 export(masked_config)
+export(masked_lp)
 export(masked_preload)
 export(masked_targets_pred)
 export(masked_tokens_tbl)
