+++ man/causal_targets_pred.Rd
@@ -1,0 +1,115 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/tr_causal.R
+\name{causal_targets_pred}
+\alias{causal_targets_pred}
+\title{Get the predictability of each element of a vector of words (or phrases) using a causal transformer}
+\usage{
+causal_targets_pred(
+  targets,
+  l_contexts = NULL,
+  log.p = getOption("pangoling.log.p"),
+  ignore_regex = "",
+  model = getOption("pangoling.causal.default"),
+  checkpoint = NULL,
+  add_special_tokens = NULL,
+  config_model = NULL,
+  config_tokenizer = NULL,
+  batch_size = 1,
+  ...
+)
+}
+\arguments{
+\item{targets}{Target words.}
+
+\item{l_contexts}{Left context for each word in \code{x}. If \code{l_contexts} is used,
+\code{by} is ignored. Set \code{by = NULL} to avoid a message notifying that.}
+
+\item{log.p}{Base of the logarithm used for the output predictability values.
+If \code{TRUE} (default), the natural logarithm (base \emph{e}) is used.
+If \code{FALSE}, the raw probabilities are returned.
+Alternatively, \code{log.p} can be set to a numeric value specifying
+the base of the logarithm (e.g., \code{2} for base-2 logarithms).}
+
+\item{ignore_regex}{Can ignore certain characters when calculates the log
+probabilities. For example \verb{^[[:punct:]]$} will ignore
+all punctuation  that stands alone in a token.}
+
+\item{model}{Name of a pre-trained model or folder.}
+
+\item{checkpoint}{Folder of a checkpoint.}
+
+\item{add_special_tokens}{Whether to include special tokens. It has the
+same default as the
+\href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoTokenizer}{AutoTokenizer} method in Python.}
+
+\item{config_model}{List with other arguments that control how the
+model from Hugging Face is accessed.}
+
+\item{config_tokenizer}{List with other arguments that control how the tokenizer from Hugging Face is accessed.}
+
+\item{batch_size}{Maximum size of the batch. Larges batches speedup
+processing but take more memory.}
+
+\item{...}{not in use.}
+}
+\value{
+A named vector of predictability values (by default the natural logarithm of the word probability).
+}
+\description{
+Get the predictability (by default the natural logarithm of the word probability) of each element of a vector of words (or phrases) given a
+vector of left contexts using a using a causal transformer model. #'
+}
+\details{
+A causal language model (also called GPT-like, auto-regressive, or decoder
+model) is a type of large language model usually used for text-generation
+that can predict the next word (or more accurately in fact token) based
+on a preceding context.
+
+If not specified, the causal model that will be used is the one set in
+specified in the global option \code{pangoling.causal.default}, this can be
+accessed via \code{getOption("pangoling.causal.default")} (by default
+"gpt2"). To change the default option
+use \code{options(pangoling.causal.default = "newcausalmodel")}.
+
+A list of possible causal models can be found in
+\href{https://huggingface.co/models?pipeline_tag=text-generation}{Hugging Face website}.
+
+Using the  \code{config_model} and \code{config_tokenizer} arguments, it's possible to
+control how the model and tokenizer from Hugging Face is accessed, see the
+Python method
+\href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoProcessor.from_pretrained}{\code{from_pretrained}}
[…]
