+++ R/tr_causal.R
@@ -102,7 +102,7 @@ causal_next_tokens_pred_tbl <- function(context,
                                         config_model = NULL,
                                         config_tokenizer = NULL) {
   if (length(unlist(context)) > 1) stop2("Only one context is allowed in this function.")
-  message_verbose("Processing using causal model '", file.path(model, checkpoint), "'...")
+  message_verbose_model(model, checkpoint)
   trf <- lang_model(model,
                     checkpoint = checkpoint,
                     task = "causal",
@@ -131,15 +131,16 @@ causal_next_tokens_pred_tbl <- function(context,
 }
 
 
-#' Get the predictability of each element of a vector of words (or phrases) using a causal transformer
+#' Get the predictability of each element of a vector of words (or phrases) in a series of texts using a causal transformer
 #'
-#' Get the predictability (by default the natural logarithm of the word probability) of each element of a vector of words (or phrases) using a causal transformer model. See the
+#' Get the predictability (by default the natural logarithm of the word probability) of each element of a vector of words (or phrases) in a series of texts using a causal transformer model. See the
 #' [online article](https://bruno.nicenboim.me/pangoling/articles/intro-gpt2.html)
 #' in pangoling website for more examples.
 #'
 #'
 #' @param x Vector of words, phrases or texts.
 #' @param by Vector that indicates how the text should be split.
+#' @param sep Character indicating how words are separated in a sentence.
 #' @param log.p Base of the logarithm used for the output predictability values.
 #'              If `TRUE` (default), the natural logarithm (base *e*) is used.
 #'              If `FALSE`, the raw probabilities are returned.
@@ -157,12 +158,32 @@ causal_next_tokens_pred_tbl <- function(context,
 #' @return A named vector of log probabilities.
 #'
 #' @examplesIf interactive()
+#' example_data <- tribble(
+#'    ~sent_n, ~word,
+#'        1,  "The",
+#'        1,  "apple",
+#'        1,  "doesn't",
+#'        1,  "fall",
+#'        1,  "far",
+#'        1,  "from",
+#'        1,  "the",
+#'        1,  "tree.",
+#'        2,  "Don't",
+#'        2,  "judge",
+#'        2,  "a",
+#'        2,  "book",
+#'        2,  "by",
+#'        2,  "its",
+#'        2,  "cover."
+#' )
 #' causal_words_pred(
-#'   x = c("The", "apple", "doesn't", "fall", "far", "from", "the", "tree."),
+#'   x = example_data$word,
+#'   by = example_data$sent_n,
 #'   model = "gpt2"
 #' )
 #' causal_words_pred(
-#'   x = c("The", "apple", "doesn't", "fall", "far", "from", "the", "tree."),
+#'   x = example_data$word,
+#'   by = example_data$sent_n,
 #'   model = "gpt2",
 #'   log.p = 1/2  # surprisal values in bits (-log2(prob) = log(prob, base = 1/2))
 #' )
@@ -171,6 +192,7 @@ causal_next_tokens_pred_tbl <- function(context,
 #' @export
 causal_words_pred <- function(x,
                               by = rep(1, length(x)),
+                              sep = " ",
                               log.p = getOption("pangoling.log.p"),
                               ignore_regex = "",
                               model = getOption("pangoling.causal.default"),
@@ -190,14 +212,11 @@ causal_words_pred <- function(x,
   }
   
   stride <- 1 # fixed for now
-  message_verbose("Processing using causal model '", file.path(model, checkpoint), "'...")
+  message_verbose_model(model, checkpoint = checkpoint)
 
   word_by_word_texts <- get_word_by_word_texts(x, by)
 
-  pasted_texts <- lapply(
-    word_by_word_texts,
-    function(word) paste0(word, collapse = " ")
[…]
