+++ R/deprecated.R
@@ -168,6 +168,7 @@ masked_lp <- function(l_contexts,
 
         word_lp(words,
           mat = m,
+          sep = " ",
           ignore_regex = ignore_regex,
           model = model,
           add_special_tokens = add_special_tokens,
@@ -357,6 +358,7 @@ causal_lp <- function(x,
       )
       word_lp(words,
         mat = mat,
+        sep = " ",
         ignore_regex = ignore_regex,
         model = model,
         add_special_tokens = add_special_tokens,
