### 2024-11-21 ac16097
treatment of special tokens
