+++ vignettes/articles/intro-gpt2.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Introduction: Using a GPT2 transformer model to get word predictability"
+title: "Using a GPT2 transformer model to get word predictability"
 bibliography: '`r system.file("REFERENCES.bib", package="pangoling")`'
 ---
 
@@ -64,10 +64,10 @@ tokenize_lst("This is different from This")
 ```
 
 
-Going back to the initial example, because `causal_next_tokens_pred_tbl` returns log (natural) probabilities, if we exponentiate them and we sum them, we should get 1:
+Going back to the initial example, because `causal_next_tokens_pred_tbl` returns by default log natural probabilities, if we exponentiate them and we sum them, we should get 1:
 
 ```{r}
-sum(exp(df_pred$lp))
+sum(exp(df_pred$pred))
 ```
 
 Because of approximation errors, this is not exactly one.
