+++ data-raw/itemsJaegeretal2014.txt
@@ -1,0 +1,385 @@
+Experiment	ItemId	Sentence	Question	Condition	Correct_Answer	Question_type	List
+RCs_SPR_ET	1	*那個*昨晚*揍了*服務生*一頓*的*顧客*聽說過*老闆*並且*記得*他。*	那個顧客聽說過老闆嗎?	a	1	1	1
+RCs_SPR_ET	1	*那個*昨晚*服務生*揍了*一頓*的*顧客*聽說過*老闆*並且*記得*他。*	那個顧客聽說過老闆嗎？	b	1	1	2
+RCs_SPR_ET	1	*老闆*聽說過*那個*昨晚*揍了*服務生*一頓*的*顧客*並且*通知了*記者。*	那個顧客聽說過老闆嗎？	c	2	1	3
+RCs_SPR_ET	1	*老闆*聽說過*那個*昨晚*服務生*揍了*一頓*的*顧客*並且*通知了*記者。*	那個顧客聽說過老闆嗎？	d	2	1	4
+RCs_SPR_ET	2	*那輛*下午*追了*摩托車*很久*的*轎車*發現了*記者*所以*停了*下來。*	摩托車追了那輛轎車嗎？	a	2	2	1
+RCs_SPR_ET	2	*那輛*下午*摩托車*追了*很久*的*轎車*發現了*記者*所以*停了*下來。*	摩托車追了那輛轎車嗎？	b	1	2	2
+RCs_SPR_ET	2	*記者*發現了*那輛*下午*追了*摩托車*很久*的*轎車*所以*停了*下來。*	摩托車追了那輛轎車嗎？	c	2	2	3
+RCs_SPR_ET	2	*記者*發現了*那輛*下午*摩托車*追了*很久*的*轎車*所以*停了*下來。*	摩托車追了那輛轎車嗎？	d	1	2	4
+RCs_SPR_ET	3	*那個*今天*打了*男孩*幾次*的*女孩*看到了*校長*所以*假裝*讀書。*	女孩打了男孩嗎？	a	1	3	1
+RCs_SPR_ET	3	*那個*今天*男孩*打了*幾次*的*女孩*看到了*校長*所以*假裝*讀書。*	女孩打了男孩嗎？	b	2	3	2
+RCs_SPR_ET	3	*校長*看到了*那個*今天*打了*男孩*幾次*的*女孩*所以*上前*詢問。*	女孩打了男孩嗎？	c	1	3	3
+RCs_SPR_ET	3	*校長*看到了*那個*今天*男孩*打了*幾次*的*女孩*所以*上前*詢問。*	女孩打了男孩嗎？	d	2	3	4
+RCs_SPR_ET	4	*那輛*當時*撞了*自行車*兩次*的*吉普車*攔住了*警察*並且*要求*調查清楚。*	警察攔住了吉普車嗎？	a	2	4	1
+RCs_SPR_ET	4	*那輛*當時*自行車*撞了*兩次*的*吉普車*攔住了*警察*並且*要求*調查清楚。*	警察攔住了吉普車嗎？	b	2	4	2
+RCs_SPR_ET	4	*警察*攔住了*那輛*當時*撞了*自行車*兩次*的*吉普車*並且*要求*調查清楚。*	警察攔住了吉普車嗎？	c	1	4	3
+RCs_SPR_ET	4	*警察*攔住了*那輛*當時*自行車*撞了*兩次*的*吉普車*並且*要求*調查清楚。*	警察攔住了吉普車嗎？	d	1	4	4
+RCs_SPR_ET	5	*那個*剛才*推了*男孩*一下*的*婦人*偷了*店員*并且*打傷了*她。*	婦人偷了店員嗎？	a	1	1	2
[…]
