+++ man/masked_config.Rd
@@ -52,6 +52,7 @@ Other masked model functions:
 \code{\link{masked_lp}()},
 \code{\link{masked_preload}()},
 \code{\link{masked_targets_pred}()},
+\code{\link{masked_tokens_pred_tbl}()},
 \code{\link{masked_tokens_tbl}()}
 }
 \concept{masked model functions}
