+++ NAMESPACE
@@ -18,6 +18,7 @@ export(masked_config)
 export(masked_lp)
 export(masked_preload)
 export(masked_targets_pred)
+export(masked_tokens_pred_tbl)
 export(masked_tokens_tbl)
 export(ntokens)
 export(perplexity_calc)
