+++ R/deprecated.R
@@ -1,25 +1,21 @@
-#' Get the possible tokens and their log probabilities for each mask in a sentence
-#'
-#' For each mask in a sentence, get the possible tokens and their log
-#' probabilities using a masked transformer.
-#'
-#' @section More examples:
-#' See the
-#' [online article](https://bruno.nicenboim.me/pangoling/articles/intro-bert.html)
-#' in pangoling website for more examples.
-#'
-#'
-#' @param masked_sentences Masked sentences.
-#' @inheritParams masked_preload
-#' @inherit masked_preload details
-#' @return A table with the masked sentences, the tokens (`token`),
-#'         log probability (`lp`), and the respective mask number (`mask_n`).
-#' @examplesIf interactive()
-#' masked_tokens_tbl("The [MASK] doesn't fall far from the tree.",
-#'   model = "bert-base-uncased"
-#' )
-#'
-#' @family masked model functions
+#' @title Deprecated functions in package \pkg{pangoling}.
+#' @description The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at \code{help("<function>-deprecated")}.
+#' @name pangoling-deprecated
+#' @keywords internal
+NULL
+
+
+#' @title Get the possible tokens and their log probabilities for each mask in a sentence
+#' @description This function is deprecated. Use `masked_tokens_pred_tbl()` instead.
+#' @name masked_tokens_tbl-deprecated
+#' @usage masked_tokens_tbl(masked_sentences, model, add_special_tokens = NULL, config_model = NULL, config_tokenizer = NULL)
+#' @seealso \code{\link{pangoling-deprecated}}
+#' @keywords internal
+NULL
+
+#' @rdname pangoling-deprecated
+#' @section \code{masked_tokens_tbl}:
+#' For \code{masked_tokens_tbl}, use \code{\link{masked_tokens_pred_tbl}}.
 #' @export
 masked_tokens_tbl <- function(masked_sentences,
                               model = getOption("pangoling.masked.default"),
@@ -70,37 +66,21 @@ masked_tokens_tbl <- function(masked_sentences,
     tidytable::relocate(mask_n, .after = tidyselect::everything())
 }
 
-#' Get the log probability of a target word (or phrase) given a left and right context
-#'
-#' Get the log probability of a vector of target words (or phrase) given a
-#' vector of left and of right contexts using a masked transformer.
-#'
-#' @section More examples:
-#' See the
-#' [online article](https://bruno.nicenboim.me/pangoling/articles/intro-bert.html)
-#' in pangoling website for more examples.
-#'
-#'
-#' @param l_contexts Left context of the target word.
-#' @param targets Target words.
-#' @param r_contexts Right context of the target word.
-#' @inheritParams masked_preload
-#' @inheritParams causal_lp
-#' @inherit masked_preload details
-#' @return A named vector of log probabilities.
-#' @examplesIf interactive()
-#' masked_lp(
-#'   l_contexts = c("The", "The"),
-#'   targets = c("apple", "pear"),
-#'   r_contexts = c(
-#'     "doesn't fall far from the tree.",
-#'     "doesn't fall far from the tree."
-#'   ),
-#'   model = "bert-base-uncased"
-#' )
-#'
-#' @family masked model functions
+
+
+#' @title Get the log probability of a target word (or phrase) given a left and right context
+#' @description This function is deprecated. Use `masked_targets_pred()` instead.
+#' @name masked_lp-deprecated
[…]
