### 2024-12-02 fb99146
failing bert things now work
