+++ R/tr_masked.R
@@ -87,48 +87,50 @@ masked_config <- function(model = getOption("pangoling.masked.default"),
 #' @family masked model functions
 #' @export
 masked_tokens_pred_tbl <- function(masked_sentences,
-                              model = getOption("pangoling.masked.default"),
-                              add_special_tokens = NULL,
-                              config_model = NULL,
-                              config_tokenizer = NULL) {
- message_verbose_model(model, checkpoint = checkpoint, causal = FALSE)
+                                   model = getOption("pangoling.masked.default"),
+                                   checkpoint = NULL,
+                                   add_special_tokens = NULL,
+                                   config_model = NULL,
+                                   config_tokenizer = NULL) {
+  message_verbose_model(model, checkpoint = checkpoint, causal = FALSE)
   tkzr <- tokenizer(model,
-    add_special_tokens = add_special_tokens,
-    config_tokenizer = config_tokenizer
-  )
+                    add_special_tokens = add_special_tokens,
+                    config_tokenizer = config_tokenizer
+                    )
   trf <- lang_model(model,
-    task = "masked",
-    config_model = config_model
-  )
+                    checkpoint = checkpoint,
+                    task = "masked",
+                    config_model = config_model
+                    )
   vocab <- get_vocab(tkzr)
   # non_batched:
   # TODO: speedup using batches
   tidytable::map_dfr(masked_sentences, function(masked_sentence) {
     masked_tensor <- encode(list(masked_sentence), tkzr,
-      add_special_tokens = add_special_tokens
-    )$input_ids
+                            add_special_tokens = add_special_tokens
+                            )$input_ids
     outputs <- trf(masked_tensor)
     mask_pos <- which(masked_tensor$tolist()[[1]] == tkzr$mask_token_id)
     logits_masks <- outputs$logits[0][mask_pos - 1] # python starts in 0
     lp <- reticulate::py_to_r(
-      torch$log_softmax(logits_masks, dim = -1L)$tolist()
-    )
+                        torch$log_softmax(logits_masks, dim = -1L)$tolist()
+                      )
     if (length(mask_pos) <= 1) lp <- list(lp) # to keep it consistent
     # names(lp) <-  1:length(lp)
     if (length(mask_pos) == 0) {
       tidytable::tidytable(
-        masked_sentence = masked_sentence,
-        token = NA,
-        lp = NA,
-        mask_n = NA
-      )
+                   masked_sentence = masked_sentence,
+                   token = NA,
+                   lp = NA,
+                   mask_n = NA
+                 )
     } else {
       lp |> tidytable::map_dfr(~
-        tidytable::tidytable(
-          masked_sentence = masked_sentence,
-          token = vocab, lp = .x
-        ) |>
-          tidytable::arrange(-lp), .id = "mask_n")
+                                 tidytable::tidytable(
+                                              masked_sentence = masked_sentence,
+                                              token = vocab, lp = .x
+                                            ) |>
+                                 tidytable::arrange(-lp), .id = "mask_n")
     }
   }) |>
     tidytable::relocate(mask_n, .after = tidyselect::everything())
@@ -166,85 +168,91 @@ masked_tokens_pred_tbl <- function(masked_sentences,
 #' @family masked model functions
 #' @export
 masked_targets_pred <- function(l_contexts,
-                      targets,
-                      r_contexts,
-                      log.p = getOption("pangoling.log.p"),
-                      ignore_regex = "",
-                      model = getOption("pangoling.masked.default"),
-                      add_special_tokens = NULL,
-                      config_model = NULL,
-                      config_tokenizer = NULL) {
+                                targets,
+                                r_contexts,
+                                log.p = getOption("pangoling.log.p"),
[…]
