### 2024-12-03 fa95880
support for hebrew
