+++ R/tr_causal.R
@@ -82,7 +82,7 @@ causal_config <- function(model = getOption("pangoling.causal.default"),
 #' [online article](https://bruno.nicenboim.me/pangoling/articles/intro-gpt2.html)
 #' in pangoling website for more examples.
 #'
-#' @param l_context The left context.
+#' @param context The context.
 #' @param decode Should it decode the tokens into readable strings? This is relevant for special characters such as accents and diacritics, which get mangled in the tokens.
 #' @inheritParams causal_preload
 #' @inherit  causal_preload details
@@ -95,7 +95,7 @@ causal_config <- function(model = getOption("pangoling.causal.default"),
 #'
 #' @family causal model functions
 #' @export
-causal_next_tokens_pred_tbl <- function(l_context,
+causal_next_tokens_pred_tbl <- function(context,
                                         log.p = getOption("pangoling.log.p"),
                                         decode = FALSE,
                                         model = getOption("pangoling.causal.default"),
@@ -103,7 +103,8 @@ causal_next_tokens_pred_tbl <- function(l_context,
                                         add_special_tokens = NULL,
                                         config_model = NULL,
                                         config_tokenizer = NULL) {
-  if (length(unlist(l_context)) > 1) stop2("Only one context is allowed in this function.")
+  if (length(unlist(context)) > 1) stop2("Only one context is allowed in this function.")
+  if(any(!is_really_string(context))) stop2("`context` needs to contain a string.")
   message_verbose_model(model, checkpoint)
   trf <- lang_model(model,
                     checkpoint = checkpoint,
@@ -116,17 +117,22 @@ causal_next_tokens_pred_tbl <- function(l_context,
                     )
 
   # no batches allowed
-  context_tensor <- encode(list(unlist(l_context)),
+  context_tensor <- encode(list(unlist(context)),
                            tkzr,
                            add_special_tokens = add_special_tokens
                            )$input_ids
   generated_outputs <- trf(context_tensor)
-  n_tokens <- length(context_tensor$tolist()[0])
+  n_tokens <- length(context_tensor$tolist()[[1]]) # was 0
   logits_next_word <- generated_outputs$logits[0][n_tokens - 1]
   l_softmax <- torch$log_softmax(logits_next_word, dim = -1L)$tolist()
   lp <- reticulate::py_to_r(l_softmax) |>
     unlist()
   vocab <- get_vocab(tkzr, decode = decode)
+  diff_words <- length(vocab) - length(lp)
+  if(diff_words > 0) {
+    warning("Tokenizer's vocabulary is longer than the model's. Some words will have NA predictability.")
+    lp <- c(lp, rep(NA, diff_words))
+    } else if(diff_words < 0) stop2("Tokenizer's vocabulary is smaller than the model's.")
   tidytable::tidytable(token = vocab,
                        pred = lp |> ln_p_change(log.p = log.p)) |>
     tidytable::arrange(-pred)
@@ -140,7 +146,7 @@ causal_next_tokens_pred_tbl <- function(l_context,
 #' in pangoling website for more examples.
 #'
 #'
-#' @param x Vector of words, phrases or texts.
+#' @param x Vector of (non-empty) words, phrases or texts.
 #' @param by Vector that indicates how the text should be split.
 #' @param sep Character indicating how words are separated in a sentence.
 #' @param log.p Base of the logarithm used for the output predictability values.
@@ -204,6 +210,7 @@ causal_words_pred <- function(x,
                               config_tokenizer = NULL,
                               batch_size = 1,
                               ...) {
[…]
