### 2024-12-03 c3e2be4
support for hebrew
