+++ man/causal_words_pred.Rd
@@ -20,7 +20,7 @@ causal_words_pred(
 )
 }
 \arguments{
-\item{x}{Vector of words, phrases or texts.}
+\item{x}{Vector of (non-empty) words, phrases or texts.}
 
 \item{by}{Vector that indicates how the text should be split.}
 
