### 2024-12-03 198772c
fixed support for hebrew
