+++ man/perplexity_calc.Rd
@@ -36,6 +36,7 @@ perplexity_calc(lprobs, log.p = TRUE)
 }
 \seealso{
 Other general functions: 
+\code{\link{install_py_pangoling}()},
 \code{\link{set_cache_folder}()}
 }
 \concept{general functions}
