+++ man/set_cache_folder.Rd
@@ -29,6 +29,7 @@ HuggingFace Transformers: \url{https://huggingface.co/transformers/index.html}
 \url{https://huggingface.co/docs/transformers/installation?highlight=transformers_cache#cache-setup}
 
 Other general functions: 
+\code{\link{install_py_pangoling}()},
 \code{\link{perplexity_calc}()}
 }
 \concept{general functions}
