### 2025-01-09 df720c7
doc and tests updates
