+++ tests/testthat/test-tr_masked.R
@@ -18,8 +18,8 @@ test_that("bert masked works", {
       masked_sentences = "The apple doesn't fall far from the [MASK].",
       model = "google/bert_uncased_L-2_H-128_A-2"
     )
-  expect_equal(colnames(mask_1), c("masked_sentence", "token", "lp", "mask_n"))
-  expect_equal(sum(exp(mask_1$lp)), 1, tolerance = 0.0001)
+  expect_equal(colnames(mask_1), c("masked_sentence", "token", "pred", "mask_n"))
+  expect_equal(sum(exp(mask_1$pred)), 1, tolerance = 0.0001)
   mask_2 <-
     masked_tokens_pred_tbl("The apple doesn't fall far from [MASK] [MASK].",
       model = "google/bert_uncased_L-2_H-128_A-2"
@@ -55,9 +55,9 @@ test_that("bert masked works", {
 test_that("bert lp for target words works", {
   skip_if_no_python_stuff()
   lw <- masked_targets_pred(
-    l_contexts = c("The", "The"),
+    prev_contexts = c("The", "The"),
     targets = c("apple", "pear"),
-    r_contexts = c(
+    after_contexts = c(
       "doesn't fall far from the tree.",
       "doesn't fall far from the tree."
     ),
@@ -67,16 +67,16 @@ test_that("bert lp for target words works", {
     model = "google/bert_uncased_L-2_H-128_A-2"
   )
   lps <- c(
-    ms[token == "apple", ]$lp,
-    ms[token == "pear", ]$lp
+    ms[token == "apple", ]$pred,
+    ms[token == "pear", ]$pred
   )
   names(lps) <- c("apple", "pear")
   expect_equal(lw, lps)
 # two tokens target
   lw <- masked_targets_pred(
-    l_contexts = c("The"),
+    prev_contexts = c("The"),
     targets = c("tasty"),
-    r_contexts = c(
+    after_contexts = c(
       "lunch."
     ),
     model = "google/bert_uncased_L-2_H-128_A-2"
@@ -85,7 +85,7 @@ ms <- masked_tokens_pred_tbl(c("The [MASK] [MASK] lunch."),
     model = "google/bert_uncased_L-2_H-128_A-2"
   )
   lps <- c(
-    ms[token == "ta" & mask_n ==1, ]$lp+   ms[token == "##sty" & mask_n ==2, ]$lp
+    ms[token == "ta" & mask_n ==1, ]$pred+   ms[token == "##sty" & mask_n ==2, ]$pred
   )
   names(lps) <- c("tasty")
   expect_equal(lw, lps)
@@ -97,9 +97,9 @@ ms <- masked_tokens_pred_tbl(c("The [MASK] [MASK] lunch."),
 test_that("bert lp for multiple target words works", {
   skip_if_no_python_stuff()
   lw <- masked_targets_pred(
-    l_contexts = c("The", "The"),
+    prev_contexts = c("The", "The"),
     targets = c("nice apple", "pretty pear"),
-    r_contexts = c(
+    after_contexts = c(
       "doesn't fall far from the tree.",
       "doesn't fall far from the tree."
     ),
@@ -109,16 +109,16 @@ test_that("bert lp for multiple target words works", {
     model = "google/bert_uncased_L-2_H-128_A-2"
   )
   lps <- c(
-    ms[token == "nice" & mask_n==1, ]$lp + ms[token == "apple" & mask_n==2, ]$lp,
-    ms[token == "pretty" & mask_n==1, ]$lp + ms[token == "pear" & mask_n==2, ]$lp
[…]
