+++ vignettes/articles/example.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Surprisal from a causal (GPT-like) model as a cognitive processing bottleneck."
+title: "Worked-Out Example of Surprisal from a Causal (GPT) Model as a Cognitive Processing Bottleneck"
 bibliography: '`r system.file("REFERENCES.bib", package="pangoling")`'
 ---
 
@@ -11,71 +11,134 @@ knitr::opts_chunk$set(
 )
 ```
 
+This vignette demonstrates how to use the `pangoling` package to estimate the effect of *surprisal* on reading times (RTs) in a psycholinguistic experiment. 
+Surprisal is a measure of processing difficulty based on the unexpectedness of a word in its context. As introduced by @levy2008, surprisal is defined as $-\log_2 P(\text{word}|\text{context})$, where $P(\text{word}|\text{context})$ is the probability of the word given its preceding context. Higher surprisal values indicate greater processing difficulty and suggest a cognitive processing bottleneck, as readers allocate more resources to processing unexpected words. @levy2008 argues that human sentence comprehension operates efficiently by leveraging probabilistic expectations. When these expectations are violated (i.e., when surprisal is high), cognitive processing slows down, which can be observed through longer reading times in reading tasks.
+
+
+Surprisal is computed here using a GPT-2-like causal language model trained on Chinese texts. We use `pangoling` to retrieve the surprisal values. The dataset comes from @jaegeretal2015, which features Chinese sentences presented word-by-word to participants during a self-paced reading experiment.
+
+
+## Data analysis
+
+### 0. Preprocessing
+
+We load the required libraries first.
 
 ```{r}
 library(pangoling)
-library(tidytable)
-library(tictoc)
+library(tidytable) # Lightweight and faster alternative to dplyr
+library(tictoc)  # Report the time that a piece of code takes
+library(brms) # Bayesian regression models in Stan
+options(mc.cores = parallel::detectCores()) # Use all the cores needed available for the Bayesian analysis by default
 ```
 
-use data from Jaeger et al 2014, Chinese sentences shown in self-paced reading to participants.  the research question and the experimental conditions don't matter to us, sentences are shown "word-by-word" (one to three characters together) to subjects using self-paced reading, explain:
+The dataset `df_jaeger14` from @jaegeretal2015 contains Chinese sentences presented word-by-word to participants. Each word has an associated reaction time (RT). (The original research question and the experimental conditions don't matter to us.)
 
 ```{r}
 data("df_jaeger14")
 df_jaeger14
 ```
 
-We get the unique words and sentences to get the predictability (we don't want to do it with the entire text because there will be repetition across subjects and the model will run several time son identical words in identical sentences).
+To avoid redundant computations, we extract unique combinations of items, conditions, and words (We don't want to run `pangoling` on the entire dataset because there will be repetition across subjects and the model will run several times on identical words in identical sentences.):
 
 ```{r}
 df_stim <- df_jaeger14 |> distinct(item, cond, word, wordn)
-df_stim |> print(n=20)
-
- df_jaeger14 |> distinct(item, cond, word, wordn, sentence) |>
-   filter(item ==1, cond =="a")
+df_stim |> print(n = 20)
 ```
 
-We preload a gpt-2 like model trained with chinese texts (see huggingface...)
+
+
[…]
