+++ man/masked_targets_pred.Rd
@@ -83,7 +83,7 @@ in pangoling website for more examples.
 }
 
 \examples{
-If interactive()
+\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 masked_targets_pred(
   prev_contexts = c("The", "The"),
   targets = c("apple", "pear"),
@@ -93,7 +93,7 @@ masked_targets_pred(
   ),
   model = "bert-base-uncased"
 )
-
+\dontshow{\}) # examplesIf}
 }
 \seealso{
 Other masked model functions: 
