+++ inst/REFERENCES.bib
@@ -58,3 +58,13 @@ author = {Lena Jäger and Zhong Chen and Qiang Li and Chien-Jer Charles Lin and 
 keywords = {Sentence processing, Relative clause, Structural expectation, Working-memory, Surprisal, Chinese},
 abstract = {Chinese relative clauses are an important test case for pitting the predictions of expectation-based accounts against those of memory-based theories. The memory-based accounts predict that object relatives are easier to process than subject relatives because, in object relatives, the distance between the relative clause verb and the head noun is shorter. By contrast, expectation-based accounts such as surprisal predict that the less frequent object relative should be harder to process. In previous studies on Chinese relative clause comprehension, local ambiguities may have rendered a comparison between relative clause types uninterpretable. We designed experimental materials in which no local ambiguities confound the comparison. We ran two experiments (self-paced reading and eye-tracking) to compare reading difficulty in subject and object relatives which were placed either in subject or object modifying position. The evidence from our studies is consistent with the predictions of expectation-based accounts but not with those of memory-based theories.}
 }
+@article{levy2008,
+  author       = {Roger Levy},
+  title        = {Expectation-based syntactic comprehension},
+  journal      = {Cognition},
+  volume       = {106},
+  number       = {3},
+  pages        = {1126--1177},
+  year         = {2008},
+  doi          = {10.1016/j.cognition.2007.05.006}
+}
