+++ DESCRIPTION
@@ -17,7 +17,7 @@ URL: https://bruno.nicenboim.me/pangoling, https://github.com/bnicenboim/pangoli
 BugReports: https://github.com/bnicenboim/pangoling/issues
 License: MIT + file LICENSE
 Encoding: UTF-8
-LazyData: false
+LazyData: true
 Imports:
     data.table,
     memoise,
@@ -39,6 +39,6 @@ RoxygenNote: 7.3.1
 Roxygen: list(markdown = TRUE)
 Depends: 
     R (>= 4.1.0)
-VignetteBuilder: knitr
+# VignetteBuilder: knitr
 StagedInstall: yes
 Language: en-US
