+++ inst/REFERENCES.bib
@@ -68,3 +68,10 @@ abstract = {Chinese relative clauses are an important test case for pitting the 
   year         = {2008},
   doi          = {10.1016/j.cognition.2007.05.006}
 }
+@book{nicenboim2025introduction,
+  title={An introduction to Bayesian data analysis for cognitive science},
+  author={Nicenboim, Bruno and Schad, Daniel and Vasishth, Shravan},
+  journal={Under contract with Chapman and Hall/CRC statistics in the social and behavioral sciences series},
+  url ={https://bruno.nicenboim.me/bayescogsci/},
+  year={2025}
+}
