### 2025-01-10 4c6ecdc
docs and added decode to tokenize
