+++ tests/testthat/test-tr_utils.R
@@ -5,7 +5,9 @@ test_that("utils works", {
   expect_true(is.character(transformer_vocab()))
   sents <- c("This is a sentence.", "This ain't a sentence.")
   expect_equal(lengths(tokenize_lst(sents)), ntokens(sents))
-})
+  expect_equal(tokenize_lst(sents, decode = TRUE), list(c("This", " is", " a", " sentence", "."), c("This", " ain", 
+                                                                                                    "'t", " a", " sentence", ".")))
+  })
 
 test_that("messages work", {
   skip_if_no_python_stuff()
