+++ man/masked_config.Rd
@@ -48,9 +48,7 @@ masked_config(model = "bert-base-uncased")
 \dontshow{\}) # examplesIf}
 }
 \seealso{
-Other masked model functions: 
-\code{\link{masked_preload}()},
-\code{\link{masked_targets_pred}()},
-\code{\link{masked_tokens_pred_tbl}()}
+Other masked model helper functions: 
+\code{\link{masked_preload}()}
 }
-\concept{masked model functions}
+\concept{masked model helper functions}
