+++ man/causal_words_pred.Rd
@@ -130,10 +130,8 @@ causal_words_pred(
 }
 \seealso{
 Other causal model functions: 
-\code{\link{causal_config}()},
 \code{\link{causal_next_tokens_pred_tbl}()},
 \code{\link{causal_pred_mats}()},
-\code{\link{causal_preload}()},
 \code{\link{causal_targets_pred}()},
 \code{\link{causal_tokens_pred_tbl}()}
 }
