+++ man/causal_next_tokens_pred_tbl.Rd
@@ -87,9 +87,7 @@ causal_next_tokens_pred_tbl(
 }
 \seealso{
 Other causal model functions: 
-\code{\link{causal_config}()},
 \code{\link{causal_pred_mats}()},
-\code{\link{causal_preload}()},
 \code{\link{causal_targets_pred}()},
 \code{\link{causal_tokens_pred_tbl}()},
 \code{\link{causal_words_pred}()}
