### 2025-01-10 #575 @bnicenboim
Phew, it took some time but I'm mostly done.

 I'm checking that I still comply with all the requisites of ropensci. Then, I'll report the changes I made and how I answered the reviewers' issues.
