+++ R/zzz.R
@@ -11,8 +11,8 @@ torch <- NULL
 
   # use superassignment to update global reference
   transformers <<- reticulate::import("transformers",
-    delay_load = TRUE,
-    convert = FALSE)
+                                      delay_load = TRUE,
+                                      convert = FALSE)
   inspect <<- reticulate::import("inspect", delay_load = TRUE, convert = TRUE) 
   torch <<- reticulate::import("torch", delay_load = TRUE, convert = FALSE)
   # TODO message or something if it's not installed
@@ -41,8 +41,14 @@ torch <- NULL
 }
 
 .onAttach <- function(libname, pkgname) {
-    packageStartupMessage(pkgname, 
-    " version ", 
-    utils::packageVersion(pkgname),
-    "\nAn introduction to the package can be found in https://bruno.nicenboim.me/pangoling/articles/\n Notice that pretrained models and tokenizers are downloaded from https://huggingface.co/ the first time they are used.\n For changing the cache folder use:\n set_cache_folder(my_new_path)")
+  packageStartupMessage(pkgname,
+                        " version ",
+                        utils::packageVersion(pkgname),
+                        "\nAn introduction to the package can be found in ",
+                        "https://bruno.nicenboim.me/pangoling/articles/\n",
+                        "Notice that pretrained models and tokenizers are ",
+                        "downloaded from https://huggingface.co/ the first ",
+                        "time they are used.\n",
+                        "For changing the cache folder use:\n",
+                        "set_cache_folder(my_new_path)")
 }
