+++ R/datasets.R
@@ -1,9 +1,9 @@
 #' Self-Paced Reading Dataset on Chinese Relative Clauses
 #'
 #' This dataset contains data from a self-paced reading experiment on Chinese
-#' relative clause comprehension. It is structured to support analysis of reaction
-#' times, comprehension accuracy, and surprisal values across various experimental
-#' conditions in a 2x2 fully crossed factorial design:
+#' relative clause comprehension. It is structured to support analysis of
+#' reaction times, comprehension accuracy, and surprisal values across various
+#' experimental conditions in a 2x2 fully crossed factorial design:
 #'
 #' - **Factor I**: Modification type (subject modification; object modification)
 #' - **Factor II**: Relative clause type (subject relative; object relative)
@@ -18,19 +18,29 @@
 #' \describe{
 #'   \item{subject}{Participant identifier, a character vector.}
 #'   \item{item}{Trial item number, an integer.}
-#'   \item{cond}{Experimental condition, a character vector indicating variations in sentence structure (e.g., "a", "b", "c", "d").}
+#'   \item{cond}{Experimental condition, a character vector indicating 
+#'   variations in sentence structure (e.g., "a", "b", "c", "d").}
 #'   \item{word}{Chinese word presented in each trial, a character vector.}
 #'   \item{wordn}{Position of the word within the sentence, an integer.}
-#'   \item{rt}{Reaction time in milliseconds for reading each word, an integer.}
-#'   \item{region}{Sentence region or phrase type (e.g., "hd1", "Det+CL"), a character vector.}
-#'   \item{question}{Comprehension question associated with the trial, a character vector.}
-#'   \item{accuracy}{Binary accuracy score for the comprehension question (1 = correct, 0 = incorrect).}
-#'   \item{correct_answer}{Expected correct answer for the comprehension question, a character vector ("Y" or "N").}
+#'   \item{rt}{Reaction time in milliseconds for reading each word, 
+#'   an integer.}
+#'   \item{region}{Sentence region or phrase type (e.g., "hd1", "Det+CL"), 
+#'   a character vector.}
+#'   \item{question}{Comprehension question associated with the trial, a 
+#'   character vector.}
+#'   \item{accuracy}{Binary accuracy score for the comprehension question 
+#'   (1 = correct, 0 = incorrect).}
+#'   \item{correct_answer}{Expected correct answer for the comprehension 
+#'   question, a character vector ("Y" or "N").}
 #'   \item{question_type}{Type of comprehension question, a character vector.}
-#'   \item{experiment}{Name of the experiment, indicating self-paced reading, a character vector.}
-#'   \item{list}{Experimental list number, for counterbalancing item presentation, an integer.}
-#'   \item{sentence}{Full sentence used in the trial with words marked for analysis, a character vector.}
-#'   \item{surprisal}{Model-derived surprisal values for each word, a numeric vector.}
+#'   \item{experiment}{Name of the experiment, indicating self-paced reading, a
+#'   character vector.}
+#'   \item{list}{Experimental list number, for counterbalancing item 
+#'   presentation, an integer.}
+#'   \item{sentence}{Full sentence used in the trial with words marked for 
+#'   analysis, a character vector.}
+#'   \item{surprisal}{Model-derived surprisal values for each word, a numeric 
+#'   vector.}
 #' }
 #'
 #' **Region codes in the dataset (column `region`)**:
@@ -38,8 +48,10 @@
 #' - **V**: Main clause verb (in object-modifications only)
 #' - **Det+CL**: Determiner+classifier
 #' - **Adv**: Adverb
[…]
