+++ tests/testthat/test-tr_masked.R
@@ -18,12 +18,13 @@ test_that("bert masked works", {
       masked_sentences = "The apple doesn't fall far from the [MASK].",
       model = "google/bert_uncased_L-2_H-128_A-2"
     )
-  expect_equal(colnames(mask_1), c("masked_sentence", "token", "pred", "mask_n"))
+  expect_equal(colnames(mask_1), 
+               c("masked_sentence", "token", "pred", "mask_n"))
   expect_equal(sum(exp(mask_1$pred)), 1, tolerance = 0.0001)
   mask_2 <-
     masked_tokens_pred_tbl("The apple doesn't fall far from [MASK] [MASK].",
-      model = "google/bert_uncased_L-2_H-128_A-2"
-    )
+                           model = "google/bert_uncased_L-2_H-128_A-2"
+                           )
   mask_2_ <-
     masked_tokens_pred_tbl(
       "[CLS] The apple doesn't fall far from [MASK] [MASK]. [SEP]",
@@ -33,8 +34,8 @@ test_that("bert masked works", {
   expect_equal(mask_2[, -1], mask_2_[, -1])
 
   nomask <- masked_tokens_pred_tbl("Don't judge a book by its cover.",
-    model = "google/bert_uncased_L-2_H-128_A-2"
-  )
+                                   model = "google/bert_uncased_L-2_H-128_A-2"
+                                   )
   masks_2_nomask <- masked_tokens_pred_tbl(
     masked_sentences = c(
       "The apple doesn't fall far from [MASK] [MASK].",
@@ -47,8 +48,8 @@ test_that("bert masked works", {
   expect_equal(
     mask_2,
     masks_2_nomask |>
-      tidytable::filter(masked_sentence ==
-        "The apple doesn't fall far from [MASK] [MASK].")
+    tidytable::filter(masked_sentence ==
+                      "The apple doesn't fall far from [MASK] [MASK].")
   )
 })
 
@@ -64,15 +65,15 @@ test_that("bert lp for target words works", {
     model = "google/bert_uncased_L-2_H-128_A-2"
   )
   ms <- masked_tokens_pred_tbl(c("The [MASK] doesn't fall far from the tree."),
-    model = "google/bert_uncased_L-2_H-128_A-2"
-  )
+                               model = "google/bert_uncased_L-2_H-128_A-2"
+                               )
   lps <- c(
     ms[token == "apple", ]$pred,
     ms[token == "pear", ]$pred
   )
   names(lps) <- c("apple", "pear")
   expect_equal(lw, lps)
-# two tokens target
+  # two tokens target
   lw <- masked_targets_pred(
     prev_contexts = c("The"),
     targets = c("tasty"),
@@ -81,11 +82,12 @@ test_that("bert lp for target words works", {
     ),
     model = "google/bert_uncased_L-2_H-128_A-2"
   )
-ms <- masked_tokens_pred_tbl(c("The [MASK] [MASK] lunch."),
-    model = "google/bert_uncased_L-2_H-128_A-2"
-  )
+  ms <- masked_tokens_pred_tbl(c("The [MASK] [MASK] lunch."),
+                               model = "google/bert_uncased_L-2_H-128_A-2"
+                               )
   lps <- c(
-    ms[token == "ta" & mask_n ==1, ]$pred+   ms[token == "##sty" & mask_n ==2, ]$pred
+    ms[token == "ta" & mask_n ==1, ]$pred + 
+    ms[token == "##sty" & mask_n ==2, ]$pred
   )
   names(lps) <- c("tasty")
[…]
