+++ tests/testthat/test-tr_causal.R
@@ -29,22 +29,28 @@ test_that("empty or small strings", {
   lp_small <- causal_words_pred(x = small_str, by = c(1, 2, 2))
   expect_equal(lp_small[1:2], c(It = NA_real_, It = NA_real_))
   expect_error(lp_small_ <- causal_words_pred(x = c("", "It"), by = c(1, 2)))
-#  expect_equal(lp_small_, c(NA_real_, "It" = NA_real_))
+  #  expect_equal(lp_small_, c(NA_real_, "It" = NA_real_))
 })
 
 if(0){
   #long inputs require too much memory
-test_that("long input work", {
-  skip_if_no_python_stuff()
-  long0 <- paste(rep("x", 1022), collapse = " ")
-  long <- paste(rep("x", 1024), collapse = " ")
-  longer <- paste(rep("x", 1025), collapse = " ")
-  lp_long0 <- causal_tokens_pred_tbl(texts = c(long0, long, longer), add_special_tokens = TRUE, batch_size = 3, model = "sshleifer/tiny-gpt2")
-  skip_on_os("windows") #the following just doesn't work on windows,
-  # but it's not that important
-    lp_long1 <- causal_tokens_pred_tbl(c(long0, long, longer), add_special_tokens = TRUE, batch_size = 1, model = "sshleifer/tiny-gpt2")
-  expect_equal(lp_long0, lp_long1)
-})
+  test_that("long input work", {
+    skip_if_no_python_stuff()
+    long0 <- paste(rep("x", 1022), collapse = " ")
+    long <- paste(rep("x", 1024), collapse = " ")
+    longer <- paste(rep("x", 1025), collapse = " ")
+    lp_long0 <- causal_tokens_pred_tbl(texts = c(long0, long, longer),
+                                       add_special_tokens = TRUE,
+                                       batch_size = 3,
+                                       model = "sshleifer/tiny-gpt2")
+    skip_on_os("windows") #the following just doesn't work on windows,
+    # but it's not that important
+    lp_long1 <- causal_tokens_pred_tbl(c(long0, long, longer),
+                                       add_special_tokens = TRUE,
+                                       batch_size = 1,
+                                       model = "sshleifer/tiny-gpt2")
+    expect_equal(lp_long0, lp_long1)
+  })
 }
 
 test_that("errors work", {
@@ -60,9 +66,10 @@ test_that("gpt2 get prob work", {
   expect_equal(cont[1]$token, "Ġtree")
   lp_prov <- causal_words_pred(x = prov_words)
   expect_equal(names(lp_prov), prov_words)
-  lp_cont <- causal_targets_pred(contexts = c("Don't judge a book by its",
-                                      "The apple doesn't fall far from the"),
-                       targets = c("cover", "tree"))
+  lp_cont <- 
+    causal_targets_pred(contexts = c("Don't judge a book by its",
+                                     "The apple doesn't fall far from the"),
+                        targets = c("cover", "tree"))
   expect_equal(lp_cont[2], lp_prov[8], tolerance = .0001)
   lp_sent2 <- causal_words_pred(x = sent2_words)
   expect_equal(names(lp_sent2), sent2_words)
@@ -112,30 +119,30 @@ test_that("gpt2 get prob work", {
     unname(lp_sent_rep[seq_along(sent_w)]),
     unname(lp_sent_rep[(length(sent_w) + 1):(2 * length(sent_w))])
   )
- df_order1 <-  data.frame(word = c(sent2_words,prov_words),
-             item = c(rep(1, each = length(sent2_words)),
-                      rep(2, each= length(prov_words))))
- df_order2 <-  data.frame(word = c(sent2_words,prov_words),
-                          item = c(rep(2, each = length(sent2_words)),
-                                   rep(1, each= length(prov_words))))
[…]

[…]
