+++ .github/workflows/pkgdown.yaml
@@ -35,6 +35,7 @@ jobs:
           extra-packages: |
             brms
             rstan
+            stan-dev/cmdstanr
             any::pkgdown
             local::.
           needs: website
