+++ R/datasets.R
@@ -86,3 +86,24 @@
 #'   group_by(region) |>
 #'   summarize(mean_rt = mean(rt, na.rm = TRUE))
 "df_jaeger14"
+
+
+#' Example dataset: Two word-by-word tokenized sentences
+#'
+#' This dataset contains tokenized words from two example sentences, split 
+#' word-by-word. It is structured to demonstrate the use of the `pangoling` 
+#' package for processing text data.
+#' package for processing text data.
+#'
+#' @format A data frame with 15 rows and 2 columns:
+#' \describe{
+#'   \item{sent_n}{(integer) Sentence number, indicating which sentence each word belongs to.}
+#'   \item{word}{(character) Tokenized words from the sentences.}
+#' }
+#' @family datasets
+#' @examples
+#' # Load the dataset
+#' data("df_sent")
+#' df_sent
+
+"df_sent"
