+++ man/masked_preload.Rd
@@ -12,7 +12,9 @@ masked_preload(
 )
 }
 \arguments{
-\item{model}{Name of a pre-trained model or folder.}
+\item{model}{Name of a pre-trained model or folder. One should be able to use
+models based on "bert". See
+\href{https://huggingface.co/models?other=bert}{hugging face website}.}
 
 \item{add_special_tokens}{Whether to include special tokens. It has the
 same default as the
