+++ R/tr_causal.R
@@ -98,7 +98,7 @@ causal_config <- function(model = getOption("pangoling.causal.default"),
 #'               relevant for special characters such as accents and 
 #'               diacritics, which get mangled in the tokens.
 #' @inheritParams causal_preload
-#' @inheritParams causal_tokens_pred_tbl
+#' @inheritParams causal_tokens_pred_lst
 #' @inherit  causal_preload details
 #' @return A table with possible next tokens and their log-probabilities.
 #' @examples
@@ -161,73 +161,90 @@ causal_next_tokens_pred_tbl <-
   }
 
 
-#' Get the predictability of each element of a vector of words (or phrases) in a
-#'  series of texts using a causal transformer
+
+#' Compute predictability using a causal transformer model
 #'
-#' Get the predictability (by default the natural logarithm of the word 
-#' probability) of each element of a vector of words (or phrases) in a series of
-#'  texts using a causal transformer model. See the
+#' These functions calculate the predictability of words or phrases using a 
+#' causal transformer model. 
+#' Depending on the use case, the following functions are available:
+#'
+#' - **`causal_targets_pred()`**: Evaluates specific target words or phrases 
+#'   based on their given contexts.
+#' - **`causal_words_pred()`**: Computes predictability for all elements of a 
+#'   vector grouped by a specified variable.
+#' - **`causal_tokens_pred_lst()`**: Computes the predictability of each token in a sentence (or group of sentences) and returns a list of results for each sentence.
+#'
+#'
+#' @details
+#' These functions calculate the predictability (by default the natural 
+#' logarithm of the word probability) of words or phrases using a causal 
+#' transformer model:
+#' 
+#' - Use `causal_targets_pred()` when you have explicit context-target pairs to evaluate, with each target word or phrase paired with a single preceding context.
+#' - Use `causal_words_pred()` when working with entire texts split into groups, such as sentences, where predictability is computed for every word or phrase in each group.
+#' - Use `causal_tokens_pred_lst()` when you want to calculate the predictability of every token in one or more sentences, returning results as a list that can be converted into a data frame if needed.
+#'
+#' See the
 #' [online article](https://bruno.nicenboim.me/pangoling/articles/intro-gpt2.html)
 #' in pangoling website for more examples.
-#'
-#'
-#' @param x Vector of (non-empty) words, phrases or texts.
-#' @param by Vector that indicates how the text should be split. By default, 
-#'           this is `" "`, but for languages, such as Chinese, which don't 
-#'           separate words, this should be set to `""`.
-#' @param sep Character indicating how words are separated in a sentence.
+#' 
+#' @param targets A character vector of target words or phrases (for `causal_targets_pred()`).
+#' @param contexts A character vector of contexts corresponding to each target (for `causal_targets_pred()`).
+#' @param x A character vector of words, phrases, or texts to evaluate (for `causal_words_pred()`).
+#' @param by A grouping variable indicating how texts are split into groups (for `causal_words_pred()`).
+#' @param sep A string specifying how words are separated within contexts or groups. Default is `" "`. For languages that don't have spaces between words (e.g., Chinese), set `sep = ""`.
+#' @param texts A vector or list of sentences or paragraphs (for `causal_tokens_pred_lst()`).
 #' @param log.p Base of the logarithm used for the output predictability values.
[…]
