+++ man/causal_next_tokens_pred_tbl.Rd
@@ -97,8 +97,6 @@ causal_next_tokens_pred_tbl(
 \seealso{
 Other causal model functions: 
 \code{\link{causal_pred_mats}()},
-\code{\link{causal_targets_pred}()},
-\code{\link{causal_tokens_pred_tbl}()},
 \code{\link{causal_words_pred}()}
 }
 \concept{causal model functions}
