+++ NAMESPACE
@@ -11,7 +11,7 @@ export(causal_pred_mats)
 export(causal_preload)
 export(causal_targets_pred)
 export(causal_tokens_lp_tbl)
-export(causal_tokens_pred_tbl)
+export(causal_tokens_pred_lst)
 export(causal_words_pred)
 export(install_py_pangoling)
 export(masked_config)

[…]
