+++ man/masked_config.Rd
@@ -45,9 +45,8 @@ for details. In case of errors check the status of
 \url{https://status.huggingface.co/}
 }
 \examples{
-\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 masked_config(model = "bert-base-uncased")
-\dontshow{\}) # examplesIf}
+
 }
 \seealso{
 Other masked model helper functions: 
