+++ man/masked_preload.Rd
@@ -55,9 +55,8 @@ for details. In case of errors check the status of
 \url{https://status.huggingface.co/}
 }
 \examples{
-\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 causal_preload(model = "bert-base-uncased")
-\dontshow{\}) # examplesIf}
+
 }
 \seealso{
 Other masked model helper functions: 
