+++ inst/REFERENCES.bib
@@ -75,3 +75,21 @@ abstract = {Chinese relative clauses are an important test case for pitting the 
   url ={https://bruno.nicenboim.me/bayescogsci/},
   year={2025}
 }
+
+@article{shannon1948mathematical,
+  title={A mathematical theory of communication},
+  author={Shannon, Claude Elwood},
+  journal={The Bell system technical journal},
+  volume={27},
+  number={3},
+  pages={379--423},
+  year={1948},
+  publisher={Nokia Bell Labs}
+}
+
+@inproceedings{hale2001probabilistic,
+  title={A probabilistic Earley parser as a psycholinguistic model},
+  author={Hale, John},
+  booktitle={Second meeting of the north american chapter of the association for computational linguistics},
+  year={2001}
+}
