+++ R/deprecated.R
@@ -45,23 +45,23 @@ masked_tokens_tbl <- function(masked_sentences,
     mask_pos <- which(masked_tensor$tolist()[[1]] == tkzr$mask_token_id)
     logits_masks <- outputs$logits[0][mask_pos - 1] # python starts in 0
     lp <- reticulate::py_to_r(
-      torch$log_softmax(logits_masks, dim = -1L)$tolist()
-    )
+                        torch$log_softmax(logits_masks, dim = -1L)$tolist()
+                      )
     if (length(mask_pos) <= 1) lp <- list(lp) # to keep it consistent
     # names(lp) <-  1:length(lp)
     if (length(mask_pos) == 0) {
       tidytable::tidytable(
-        masked_sentence = masked_sentence,
-        token = NA,
-        lp = NA,
-        mask_n = NA
-      )
+                   masked_sentence = masked_sentence,
+                   token = NA,
+                   lp = NA,
+                   mask_n = NA
+                 )
     } else {
       lp |> tidytable::map_dfr(~
                                  tidytable::tidytable(
-                                   masked_sentence = masked_sentence,
-                                   token = vocab, lp = .x
-                                 ) |>
+                                              masked_sentence = masked_sentence,
+                                              token = vocab, lp = .x
+                                            ) |>
                                  tidytable::arrange(-lp), .id = "mask_n")
     }
   }) |>
@@ -93,76 +93,14 @@ masked_lp <- function(l_contexts,
                       config_model = NULL,
                       config_tokenizer = NULL) {
   .Deprecated(new = "masked_targets_pred()")
-  stride <- 1
-  tkzr <- tokenizer(model,
-                    add_special_tokens = add_special_tokens,
-                    config_tokenizer = config_tokenizer
-                    )
-  trf <- lang_model(model,
-                    task = "masked",
-                    config_model = config_model
-                    )
-
-  message_verbose("Processing using masked model '", model, "'...")
-
-  target_tokens <- char_to_token(targets, tkzr)
-  masked_sentences <- tidytable::pmap_chr(
-    list(
-      l_contexts,
-      target_tokens,
-      r_contexts
-    ),
-    function(l, target, r) {
-      paste0(
-        l,
-        " ",
-        paste0(rep(tkzr$mask_token, length(target)), collapse = ""),
-        " ",
-        r
-      )
-    }
-  )
-
-  # named tensor list:
-  tensors_lst <- tidytable::map2(masked_sentences, targets, function(t, w) {
-    l <- create_tensor_lst(t,
-                           tkzr,
-                           add_special_tokens = add_special_tokens,
-                           stride = stride
-                           )
-    names(l) <- w
-    l
-  })
-
-  out <- tidytable::pmap(
-    list(targets, l_contexts, r_contexts, tensors_lst),
-    function(words, l, r, tensor_lst) {
-      # TODO: make it by batches
-      ls_mat <- masked_lp_mat(lapply(tensor_lst, function(t) t$input_ids),
-                              trf = trf,
-                              tkzr = tkzr,
-                              add_special_tokens = add_special_tokens,
-                              stride = stride
-                              )
-      text <- paste0(words, collapse = " ")
-      tokens <- char_to_token(text, tkzr)[[1]]
-      lapply(ls_mat, function(m) {
-        # m <- ls_mat[[1]]
-        message_verbose(l, " [", words, "] ", r)
-
-        word_lp(words,
-                mat = m,
-                sep = " ",
-                ignore_regex = ignore_regex,
-                model = model,
-                add_special_tokens = add_special_tokens,
-                config_tokenizer = config_tokenizer
[…]
