### 2025-01-19 da17852
docs and Mac error
