+++ .github/workflows/R-CMD-check.yaml
@@ -61,7 +61,8 @@ jobs:
           path_to_python <- virtualenv_python("r-reticulate")
           writeLines(sprintf("RETICULATE_PYTHON=%s", path_to_python),
                      Sys.getenv("GITHUB_ENV"))
-          py_run_string("import os; os.environ['KMP_DUPLICATE_LIB_OK']='True'")  #Error #15: Initializing libomp.dylib, but found libomp.dylib already initialized
+          writeLines(sprintf("KMP_DUPLICATE_LIB_OK = TRUE", path_to_python),
+                     Sys.getenv("GITHUB_ENV")) #Error #15: Initializing libomp.dylib, but found libomp.dylib already initialized
 
 
 
