+++ .github/workflows/R-CMD-check.yaml
@@ -57,7 +57,7 @@ jobs:
           library(reticulate)
           virtualenv_create("r-reticulate", Sys.which("python"))
           virtualenv_install("r-reticulate", python_packages)
-
+          virtualenv_remove('r-reticulate', packages = c('mkl'))
           path_to_python <- virtualenv_python("r-reticulate")
           writeLines(sprintf("RETICULATE_PYTHON=%s", path_to_python),
                      Sys.getenv("GITHUB_ENV"))
