+++ pkgdown.yml
@@ -1,12 +1,12 @@
-pandoc: 2.19.2
-pkgdown: 2.0.7
+pandoc: 3.1.11
+pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
-  intro-bert: intro-bert.html
-  intro-gpt2: intro-gpt2.html
-  troubleshooting: troubleshooting.html
-last_built: 2023-11-04T13:51Z
+  articles/example: example.html
+  articles/intro-bert: intro-bert.html
+  articles/intro-gpt2: intro-gpt2.html
+  articles/troubleshooting: troubleshooting.html
+last_built: 2025-01-19T21:42Z
 urls:
   reference: http://bruno.nicenboim.me/pangoling/reference
   article: http://bruno.nicenboim.me/pangoling/articles
-
