+++ .github/workflows/R-CMD-check.yaml
@@ -57,11 +57,11 @@ jobs:
           library(reticulate)
           virtualenv_create("r-reticulate", Sys.which("python"))
           virtualenv_install("r-reticulate", python_packages)
-          virtualenv_remove('r-reticulate', packages = c('mkl'))
+          virtualenv_remove('r-reticulate', packages = c('mkl-service'))
           path_to_python <- virtualenv_python("r-reticulate")
           writeLines(sprintf("RETICULATE_PYTHON=%s", path_to_python),
                      Sys.getenv("GITHUB_ENV"))
-          writeLines(sprintf("KMP_DUPLICATE_LIB_OK = TRUE", path_to_python),
+          writeLines("KMP_DUPLICATE_LIB_OK = TRUE",
                      Sys.getenv("GITHUB_ENV")) #Error #15: Initializing libomp.dylib, but found libomp.dylib already initialized
 
 
