### 2025-01-20 #575 @bnicenboim
Hi, I'm not sure what to do about the errors. I don't get the warnings or notes here:
https://github.com/bnicenboim/pangoling/actions/runs/12864039448/job/35861748929

It might be that  rmdcheck doesn't install the python dependencies needed? But I'm following the latest advice from reticulate (https://rstudio.github.io/reticulate/articles/package.html) in both creating a function to install the python dependencies rather than doing it automatically and how to test them using Github Actions
