+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   articles/intro-bert: intro-bert.html
   articles/intro-gpt2: intro-gpt2.html
   articles/troubleshooting: troubleshooting.html
-last_built: 2025-01-20T08:41Z
+last_built: 2025-01-20T10:35Z
 urls:
   reference: http://bruno.nicenboim.me/pangoling/reference
   article: http://bruno.nicenboim.me/pangoling/articles
