### 2025-01-20 #575 @bnicenboim
Hi everyone, after a long time, I have an update! I made a lot of changes—see the [NEWS file](https://github.com/bnicenboim/pangoling/blob/main/NEWS.md).  Renamed most functions, added more arguments, simplified the installation, added a [troubleshooting vignette](https://bruno.nicenboim.me/pangoling/articles/troubleshooting.html) and a [worked out example with Chinese sentences and surprisal values](https://bruno.nicenboim.me/pangoling/articles/example.html).

Thanks again for the reviews! The long delay turned out to be beneficial, as looking at it with fresh eyes helped me improve the documentation.

I address the reviewers' comments below.

---

## @utkuturk's Review

### Clarity issue: Python dependencies
@utkuturk mentioned difficulties with installing the Python dependencies.

**Response**: I updated the installation process to follow the latest `reticulate` recommendations. There's now an [`install_py_pangoling()`](https://bruno.nicenboim.me/pangoling/reference/install_py_pangoling.html) function to handle installation, along with a [troubleshooting vignette](https://bruno.nicenboim.me/pangoling/articles/troubleshooting.html).

### Clarity issue: Downloading models
@utkuturk suggested making it clearer that the package downloads models locally and does not operate through an API.

**Response**: This is now explicitly stated in the startup message:
```
pangoling version 0.0.0.9010
An introduction to the package can be found at https://bruno.nicenboim.me/pangoling/articles/
Notice that pretrained models and tokenizers are downloaded from https://huggingface.co/ the first time they are used.
To change the cache folder, use:
set_cache_folder(my_new_path)
```

### Advanced Configuration: `NULL` Arguments
@utkuturk noted that the documentation lacks information on how `NULL` arguments work in exported functions.

**Response**: I think this is out of scope. The package is fully functional without users needing to configure models. However, I’ve ensured that advanced users can still modify configurations if needed.

---

## @lisalevinson's Review

### Workflow: Extracting probabilities for all words
@lisalevinson typically gathers probabilities for all words in a sentence rather than just one target word and found the argument structure cumbersome, especially for masked models.

**Response**: The challenge with a position-based approach (e.g., `target = 8`) is defining what "8" refers to—words (what counts as a word?) or tokens? Instead, here’s how it's done now:

```r
(preds <- causal_tokens_pred_lst("The apple doesn't fall far from the tree."))
```
```
[[1]]
          The        Ġapple        Ġdoesn            't         Ġfall          Ġfar
           NA -1.090052e+01 -5.499081e+00 -8.279233e-04 -3.597742e+00 -2.911830e+00
        Ġfrom          Ġthe         Ġtree             .
-7.454723e-01 -2.066519e-01 -2.808024e-01 -1.300929e+00
```
```r
# this creates lists because it allows for several sentences
preds[[1]][9]  # Probability for "tree"
```
For masked models one needs to explicitly cut the sentence where the targets are and propose the possible targets:

```r
masked_targets_pred(
  prev_contexts = c("The", "The"),
  targets = c("apple", "pear"),
  after_contexts = c("doesn't fall far from the tree.", "doesn't fall far from the tree."),
  model = "bert-base-uncased"
)
```

### Reporting and documentation of probabilities
@lisalevinson noted that the log probabilities are in natural log but couldn't find this documented. Since base-2 logs (bits) are common in psycholinguistics, it would be useful to specify the log base.

**Response**: This was a great suggestion! I added a `log.p` argument:
```
Base of the logarithm used for the output predictability values. If TRUE (default), the natural logarithm (base e)
is used. If FALSE, the raw probabilities are returned. Alternatively, log.p can be set to a numeric value specifying
the base of the logarithm (e.g., 2 for base-2 logarithms). To get surprisal in bits (rather than predictability), set log.p = 1/2.
```

### Output formats
@lisalevinson found the function outputs inconsistent and suggested a vignette on reshaping the data for analysis.

**Response**: I standardized the outputs as much as possible and added a [vignette](https://bruno.nicenboim.me/pangoling/articles/example.html) on using `pangoling` in psycholinguistic research.

### Non-English usage and examples
@lisalevinson suggested including examples for non-English languages, particularly those without spaces (e.g., Chinese).

**Response**: I added a `sep` argument to handle languages without spaces:
```
A string specifying how words are separated within contexts or groups. Default is " ".
For languages that don't have spaces between words (e.g., Chinese), set sep = "".
```

Also the [vignette](https://bruno.nicenboim.me/pangoling/articles/example.html) is with an example in Chinese and using surprisal rather than predictability ;)

Multi-token words are already handled correctly—the package sums log probabilities across tokens.

---

## Specific Function Comments

### `causal_lp()`
@lisalevinson found it confusing that `x` could be a full sentence or a set of targets.

**Response**: This is now split into two functions:
- [`causal_words_pred()`](https://bruno.nicenboim.me/pangoling/reference/causal_predictability.html) for word-by-word (or phrase-by-phrase) full sentences.  This one uses `x`
- [`causal_targets_pred()`](https://bruno.nicenboim.me/pangoling/reference/causal_predictability.html) for target words.  This one uses `targets`.

### `batch_size` Documentation
@lisalevinson found the explanation unclear.

**Response**: I clarified the documentation:
> "batch_size: Maximum batch size. Larger batches speed up processing but require more memory."

### `causal_lp_mats()` (now `causal_pred_mats()`)
@lisalevinson suggested clarifying why it returns a matrix instead of a dataframe.

**Response**: It returns a matrix due to memory constraints and because all values are numeric. The documentation now explicitly states:

```
The function splits the input x into groups specified by the by argument and processes each group independently. For each group, the model computes the predictability of each token in its vocabulary based on preceding context.

Each matrix contains:
- Rows representing the model's vocabulary.
- Columns corresponding to tokens in the group (e.g., a sentence or paragraph).
- By default, values in the matrices are the natural logarithm of word probabilities.

```

She also suggested including examples for multiple sentences and how to wrangle the output.

**Response**: I added examples for processing multiple sentences.

### Token vs. word Confusion
@lisalevinson pointed out that column names in some functions refer to "words," but they are actually tokens (e.g., `"doesn"` and `"’t"` are separate).

**Response**: I corrected the documentation to consistently refer to "tokens" instead of "words" where applicable. The function descriptions now clarify when they operate on tokenized text rather than full words.

### `causal_tokens_lp_tbl()` (now `causal_tokens_pred_lst()`)
@lisalevinson found the argument name `.id` unintuitive.

**Response**: `.id` has been removed—it’s no longer needed.

She also found the distinction between `texts` (full sentences) and `x` (target words) inconsistent across functions.

**Response**: `causal_tokens_pred_lst()` still takes `texts` because it processes full sentences or paragraphs and returns token probabilities. In contrast, `causal_words_pred()`  take `x`, which represents smaller linguistic units like words or phrases. This distinction is now explicitly explained in the [documentation](https://bruno.nicenboim.me/pangoling/reference/causal_predictability.html).

### `masked_lp()` preferring a different argument structure.

**Response**: I answered also above.

### `masked_tokens_tbl()`
@lisalevinson suggested a clearer argument structure and better differentiation from `causal_tokens_lp_tbl()`.

**Response**:
- `masked_tokens_tbl()` is now [`masked_tokens_pred_tbl()`](https://bruno.nicenboim.me/pangoling/reference/masked_predictability.html).

**Response**: The difference comes from how masked models work—masked predictions return probabilities for a single masked token position, whereas causal models generate predictions for all possible next tokens based on the context.  I updated the documentation to clarify this distinction.

For masked models, this follows the standard Hugging Face approach:
```r
masked_tokens_pred_tbl("The [MASK] doesn't fall far from the tree.", model = "bert-base-uncased")
```
This returns all possible completions with their log probabilities.

### Test Failures:
**Response**: All tests pass.

---

This should now be clearer and more consistent. I tried to summarize the points to be able to answer them, if something is missing (or I misunderstood it), please let me know.
