### 2025-01-22 #575 @emilyriederer
Thank you so much for the detailed notes and updates, @bnicenboim !

@lisalevinson, @utkuturk - I know it has been a bit since you completed your reviews. Could you please reference @bnicenboim 's update comments above and assess if you feel these address the themes you raised? Ideally, you could please acknowledge using [this template](https://devguide.ropensci.org/approval2template.html)
