### 2025-02-02 #575 @maurolepore
Dear @bnicenboim this is to mark the start of my EiC rotation. I'm reviewing all open issues and noting what I see.

I see @bnicenboim addressed reviewers' requests, @emilyriederer asked reviewers for approvals and @utkuturk already approved. We're waiting for @lisalevinson to approve or request further changes.
