### 2025-02-19 #575 @lisalevinson
## Reviewer Response

Thanks for this new revised version - it is much improved! I approve of this revision, but I have a couple of minor comments/questions for consideration at some point.

For the batch size in `causal_word_pred()`, I don't understand what the batch is from the documentation - batches of the groups/sentences, or batches of tokens/words? In the Mandarin example, I switched the batch between size 1 to 10 - the former had tokens per batch around 27-33 tokens (characters), which makes sense as sentence per batch, but the latter said batches all said they had 33 tokens. I don't follow what that means. Not sure if something is weird with the output message or the batches themselves.

Another concern is about the word/token order within groups/sentences. The word context/prefix is based purely on row order? It doesn't seem like the wordn value is used at all. This seems fragile if tokens possibly get sorted/rearranged, and it might not be an obvious problem if one is not carefully inspecting the order of tokens in the outputs. I know usually it will be fine but I always am nervous when anything in data analysis crucially depends on row order.

#### Final approval (post-review)

Estimated hours spent reviewing: 13
