### 2025-02-28 #575 @bnicenboim
Hi, Thanks a @utkuturk and @lisalevinson !!
Now some questions for  @emilyriederer and @maurolepore.
What is the next step in following up on this? Would it be possible to publish something on the JOSS based on the package? What's the role of the review on this?
