### 2025-02-28 #575 @maurolepore
@emilyriederer I hope you don't mind ... I'll jump in while I'm doing an overall review of issues.

@bnicenboim, I see both reviewers approved but @lisalevinson seems to have some thoughtful comments. I'll go ahead and call the bot to approve the package, but please so acknowledge @lisalevinson comments and share your thoughts.

RE JOSS, our guidelines suggest your next step is to reach out to them, letting them know of our approval:

> If you feel your package is in scope for the [Journal of Open-Source Software](https://joss.theoj.org/) (JOSS), do not submit it to JOSS consideration until after the rOpenSci review process is over: if your package is deemed in scope by JOSS editors, only the accompanying short paper would be reviewed, (not the software that will have been extended reviewed by rOpenSci by that time). Not all rOpenSci packages will meet the criteria for JOSS.
-- https://devdevguide.netlify.app/softwarereview_author.html#preparing-for-submission

Also the bot will reply with a number of tasks. Please complete them ASAP and do let us know if you need help.
