### 2025-02-28 #575 @ropensci-review-bot
Approved! Thanks @bnicenboim for submitting and @lisalevinson, @utkuturk for your reviews! :grin:
