### 2025-03-10 #575 @bnicenboim
Hi again @emilyriederer and @maurolepore,

Ok, I discovered that I can find my package here and see it failing:
https://github.com/r-universe/ropensci/

At least for pkgdwn, the issue is that I installed a bunch of extra R and Python packages to build the website. Now, I'm relying on the yml of ropensci, which is not doing that. How should I fix this?

It also fails when it builds the binaries because it tries to run the examples, which also require Python scripts. But I guess I can fix that on my end, not allowing to run the examples when python packages are not installed with @example If.  (And I guess I'll need that submit as cran as well).

In any case, the first issue (pkgdown website) remains. Thanks!
