### 2025-03-11 #575 @bnicenboim
ok, I managed. I had to do this: https://ropensci.org/blog/2019/12/08/precompute-vignettes/
