### 2025-03-12 #575 @bnicenboim
Hi @ropensci/blog-editors,

I definitely want to write a blog post. I'm currently trying to submit the package to CRAN; perhaps I can write about it once I successfully get it approved.

Also, the transfer didn't go as smoothly as I had hoped. There’s some missing information scattered around. Should I open an issue here: https://github.com/ropensci/dev_guide?
