### 2025-03-12 #575 @mpadge
@bnicenboim The blog editors will get back to you about that question, and I'm just responding to the transfer issues. I see both [repo](https://github.com/ropensci/pangoling/) and [docs](https://docs.ropensci.org/pangoling/) have been transferred and built as expected, so am unsure what you mean by not going smoothly. Could you briefly describe here, and we'll be able to judge whether there may be an issue for the dev guide? Thanks, and congrats on finalizing the review process, and on a great package!
