### 2025-03-12 #575 @bnicenboim
Hi, yes, it built as expected after a lot of tinkering.

1. I followed https://devguide.ropensci.org/redirect.html, but the information there wasn't accurate for my case. I'm not sure how non-standard my website is, but I had to add the "pangoling/" directory inside the `docs/` folder.

2. The pkgdown website wasn’t building, and I had no idea why. Following https://docs.ropensci.org/pangoling/ led to a 404 error. After a while, I discovered https://github.com/r-universe/ropensci/, and then spent some more time searching to find https://ropensci.org/blog/2019/12/08/precompute-vignettes/. I didn't know that I couldn't use Python in my pkgdown configuration (I had added all the dependencies and Python software I wanted a long time ago to my yml).  Also, the post I referenced is missing some important details. It's true that the images need to be moved, but the `fig.path` in `knitr::opts_chunk$set()` also needs to be adjusted. That is  because vignettes should not rely on folder structures; otherwise, I would receive a note about it. In addition, I was using `system.file("REFERENCES.bib", package="pangoling")` in my vignettes, which created a hardcoded path that led to additional problems.
