### 2025-03-12 #575 @bnicenboim
And one last thing about the transfer: Most of my examples in the documentation rely on Python. It would be wonderful if there were an option to include Python dependencies in my pkgdown site, even if it is hosted on ropensci. Of course, I understand that this may not be feasible.
