### 2025-03-13 #575 @bnicenboim
The python packages (torch and transformers) are needed for everything :) But I precompute my vignettes now, so the vignettes are fine. Now it's mostly for the examples in the documentation.
Reticulate [last recomendations](https://rstudio.github.io/reticulate/reference/py_require.html) are not add the python packages in the DESCRIPTION FILE.

How would I run you install_py_pangoling() on r-universe?  Should I check if MY_UNIVERSE==TRUE when I load the package?  Or is there another way?
