### 2025-03-13 #575 @bnicenboim
I see now that the latest version of reticulate allows you to have packages in an ephemeral environment, maybe that would be the solution. I need to investigate this further....
