### 2025-03-13 #575 @steffilazerte
Hi @bnicenboim,

(Sorry to interrupt the conversation!)

I'm excited to hear that you're interested in writing a blog post for us!

We have two types of blog posts: either [an introduction to your package featuring technical details](https://ropensci.org/tech-notes/) for a technical audience or [a longer post with some narrative](https://ropensci.org/blog/) about its development or something you learned, and possibly an example of its use for a broader readership. The difference between a technote and a blog post is more about the style and content then length, but technical posts are often shorter than blog posts (no requirement either way).

To contribute a post, we'll have you open a PR to the roweb3 repository. About a week or two before our publication date, you can ping me and I'll review the post and give you some suggestions for content and style. You'll consider my feedback, make any changes you'd like to make, and then we merge and publish! We have a flexible timeline right now (with the exception that I'll be unavailable to review for a couple of weeks in mid-April), so whenever you're ready is good for us. However, if you work better with a deadline, we can do that too.

You can read more about the details of writing an rOpenSci blog post here: https://blogguide.ropensci.org/ (and don't hesitate to ask if you have any questions).

You can reach me on the Slack (Steffi LaZerte; you should have gotten an invitation by now?) or via email ([sel@steffilazerte.ca](mailto:sel@steffilazerte.ca)), and/or ping me when you open the PR to roweb3 (@steffilazerte).

Thanks!
